% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs-spectrum.R
\name{add_dec31_art}
\alias{add_dec31_art}
\title{Disaggregate the number on ART Dec 31 to single age}
\usage{
add_dec31_art(spec, pjnz)
}
\description{
Disaggregate the number on ART Dec 31 to single age
}
