% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report-plots.R
\name{age_bar_plotly}
\alias{age_bar_plotly}
\title{Plotly barplot comparing age distribution of survey inputs and naomi outputs}
\usage{
age_bar_plotly(df, ind, quarter)
}
\arguments{
\item{df}{Inputs_outputs dataframe containing matched model estimates and data inputs.}

\item{ind}{Indicator filter.}

\item{quarter}{Calendar quarter filter.}
}
\description{
Plotly barplot comparing age distribution of survey inputs and naomi outputs
}
