% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-time-series.R
\name{aggregate_anc}
\alias{aggregate_anc}
\title{Aggregate ANC data according to area hierarchy}
\usage{
aggregate_anc(anc, shape)
}
\arguments{
\item{anc}{Path to file containing ANC data or ANC data object}

\item{shape}{Path to file containing geojson areas data or areas data object}
}
\value{
Aggregated ANC data containing columns area_id, area_name, area_level,
area_level_label, sex,age_group, time_period, year, quarter, calendar_quarter,
anc_clients, anc_known_neg, anc_already_art, anc_tested and anc_tested_pos,
births_clients_ratio
}
\description{
Take ANC and shape file paths or data files and aggregate
art_current according to area hierarchy provided
}
