% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-time-series.R
\name{aggregate_art}
\alias{aggregate_art}
\title{Aggregate ART data according to area hierarchy}
\usage{
aggregate_art(art, shape)
}
\arguments{
\item{art}{Path to file containing ART data or ART data object}

\item{shape}{Path to file containing geojson areas data or areas data object}
}
\value{
Aggregated ART data containing columns area_id, area_name,
area_level, area_level_label, parent_area_id, sex, age_group, time_period,
year, quarter,calendar_quarter and art_current
}
\description{
Take ART and shape file paths or files and aggregate
art_current according to area hierarchy provided
}
