% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agyw-integration.R
\name{agyw_adjust_sexbehav_msm_pwid}
\alias{agyw_adjust_sexbehav_msm_pwid}
\title{Adjust male sexual behavior risk groups by MSM + PWID proportions}
\usage{
agyw_adjust_sexbehav_msm_pwid(outputs, options, msm_est, pwid_est)
}
\arguments{
\item{outputs}{Naomi output.}

\item{options}{Naomi model options.}

\item{msm_est}{5-year estimates of MSM PSEs generated from \code{agyw__disaggregate_msm()}.}

\item{pwid_est}{5-year estimates of MSM PSEs generated from \code{agyw__disaggregate_pwid()}.}

\item{sae_path}{Path to female estimates of sexual behavior risk group.}
}
\value{
District level estimates of male sexual risk behaviour groups

Estimates are generated for the following groups:
\itemize{
\item \code{nosex12m}:
\item \code{sexcohab}:
\item \code{sexonregplus}:
\item \code{sexonreg}:
\item \code{msm}:
\item \code{pwid}:
}

Calculation steps:
\enumerate{
\item Align admin0/admin1 MSM and PWID proportions with SRB SAE estimates.
\item Subtracting MSM and PWID proportionally from all SRB groups.
}

Load male SRB proportions
Subtracting MSM and PWID proportionally from all SRB risk groups
(FSW was just from high-risk females)
}
\description{
Adjust male sexual behavior risk groups by MSM + PWID proportions
}
