% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agyw-integration.R
\name{agyw_calculate_incidence_female}
\alias{agyw_calculate_incidence_female}
\title{Calculate incidence for female SRB groups.}
\usage{
agyw_calculate_incidence_female(
  naomi_output,
  options,
  female_srb,
  female_logit_prevalence,
  survey_year = 2018
)
}
\arguments{
\item{naomi_output}{Naomi output.}

\item{options}{Naomi model options.}

\item{female_srb}{FSW adjusted estimates of female sexual risk groups generated by \code{agyw_adjust_sexbehav_fsw()}.}

\item{female_logit_prevalence}{Risk adjusted estimates of female prevalence in sexual risk groups generated by \code{agyw_calculate_prevalence_female()}.}

\item{survey_year}{Survey year to sample from the SAE model. Default is 2018. Survey year should be updated to most current household survey in the country - for countries without recent household surveys, leave at 2018 - the spatiotemporal
model of sexual behaviour fitted to all countries has the most data for in roughly 2018.}
}
\value{
Wide format output required for the AGYW workbook.
}
\description{
While maintaining age/sex/district-specific HIV incidence from Naomi, distribute
HIV incidence between our 4 different behavioural groups utilizing IRRs from the
literature
}
\keyword{internal}
