% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agyw-integration.R
\name{agyw_calculate_incidence_male}
\alias{agyw_calculate_incidence_male}
\title{Calculate incidence in high risk male key populations}
\usage{
agyw_calculate_incidence_male(
  naomi_output,
  options,
  male_srb,
  male_logit_prevalence,
  survey_year = 2018
)
}
\arguments{
\item{options}{Naomi options extracted from outputs}

\item{male_srb}{Estimates of male sexual risk groups generated by \code{agyw_adjust_sexbehav_msm_pwid()}}

\item{male_logit_prevalence}{Risk adjusted estimates of male prevalence in sexual risk groups generated by \code{agyw_calculate_prevalence_male()}}

\item{survey_year}{NOTES:: add in description + when this should be adjusted. Hardcoded to 2018.}

\item{outputs}{Naomi output.}
}
\value{
Wide format output required for the AGYW workbook
}
\description{
Calculate incidence in high risk male key populations
}
