% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agyw-integration.R
\name{agyw_calculate_prevalence_female}
\alias{agyw_calculate_prevalence_female}
\title{Calculate prevalence for female SRB groups.}
\usage{
agyw_calculate_prevalence_female(
  naomi_output,
  options,
  fsw_est,
  female_srb,
  survey_year_sample = 2018
)
}
\arguments{
\item{options}{Naomi model options.}

\item{fsw_est}{5-year estimates of MSM PSEs generated from \code{agyw_disaggregate_fse()}.}

\item{outputs}{Naomi output.}

\item{female_sexbehav}{KP adjusted estimates of female SRB groups generated by \code{agyw_adjust_sexbehav_fsw()}}

\item{female_hiv_path}{Path to SRB HIV estimates from household surveys (last updated XX-XX-XX).}

\item{pse_path}{Path to KP PSEs last updated (XX-XX-XX).}

\item{survey_year}{Year of survey to sample estimates.}
}
\value{
SRB PSEs with logit prevalence estimates.

To calculate district-age-sex-sexual behaviour-specific HIV prevalence, we maintain
HIV prevalence from Naomi for a district-age-sex, but disaggregate to different
risk behaviours using:
(1) HIV prevalence ratios from household surveys for those reporting no
sex vs one cohabiting vs non-regular sexual partner(s), and
(2) a linear regression through admin-1 level estimates of the ratio of KP
prevalence to gen-pop prevalence used to predict an age-district-specific
FSW to general population prevalence ratio.

Naomi estimates of PLHIV and population by district and age band
Extract country specific national FSW prevalence
THIS IS NOW USING SINGLE COUNTRY INSTEAD OF ALL COUNTRIES
KP regression: FSW prevalence relative to general prevalence
########## THIS REGRESSION SHOULD BE TAKING DATA FROM ALL ADMIN-1 LEVEL
Modelled estimates of proportion in each risk group
Calculate prevalence in each category
Log odds ratio from SRB group survey prevalence
Calculate logit prevalence and format
}
\description{
Calculate prevalence for female SRB groups.
}
