% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agyw-integration.R
\name{agyw_disaggregate_msm}
\alias{agyw_disaggregate_msm}
\title{Disaggregate admin1 MSM proportions from Oli's KP model to 5-age groups}
\usage{
agyw_disaggregate_msm(outputs, options, naomi_pop, kp_consensus)
}
\arguments{
\item{outputs}{Naomi output.}

\item{options}{Naomi model options.}

\item{kp_consensus}{Key pop consensus estimates.}

\item{naomi_population}{Naomi population estimates for T2.}
}
\value{
District level MSM estimates by 5-year age bands for ages 15-49.
Extract country specific national MSM PSEs
Check for consensus estimate of MSM
MSM age distribution parameters in ZAF from Thembisa
Downloaded from: https://www.thembisa.org/content/downloadPage/Thembisa4_3report
Distribution function of the gamma
Calculate proportion of sexually active population using Kinh's country specific
estimates of age at first sex and naomi population
Select birth cohort from 2000, to turn 15 in 2015
Calculate sexually active population by age and sex for each district
Adjusting country specific sexual debut estimates with age distribution of
MSM from Thembisa
}
\description{
Disaggregate admin1 MSM proportions from Oli's KP model to 5-age groups
}
