% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agyw-integration.R
\name{agyw_format_naomi}
\alias{agyw_format_naomi}
\title{Format naomi outputs for PSE tool}
\usage{
agyw_format_naomi(outputs, options)
}
\arguments{
\item{outputs}{Naomi output}

\item{options}{Naomi model options.}
}
\value{
Naomi indicators formatted for the AGYW workbook.
}
\description{
Format naomi outputs for PSE tool
}
