% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agyw-integration.R
\name{agyw_generate_risk_populations}
\alias{agyw_generate_risk_populations}
\title{Generate outputs to update AGYW tool.}
\usage{
agyw_generate_risk_populations(naomi_output, pjnz, survey_year = 2018)
}
\arguments{
\item{naomi_output}{Path to naomi output (zip file or hintr object).}

\item{pjnz}{Path to spectrum file.}

\item{survey_year}{Survey year to sample from the SAE model. Default is 2018. Survey year should be updated to most current household survey in the country - for countries without recent household surveys, leave at 2018 - the spatiotemporal
model of sexual behaviour fitted to all countries has the most data for in roughly 2018.}

\item{male_srb}{Estimates of male sexual risk groups generated by \code{agyw_adjust_sexbehav_msm_pwid()}}

\item{male_logit_prevalence}{Risk adjusted estimates of male prevalence in sexual risk groups generated by \code{agyw_calculate_prevalence_male()}}
}
\value{
Output files to update AGYW excel workbook.
}
\description{
Generate outputs to update AGYW tool.
}
\keyword{internal}
