% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputs.R
\name{align_inputs_outputs}
\alias{align_inputs_outputs}
\title{Align model data inputs and model estimates}
\usage{
align_inputs_outputs(naomi_data, indicators, meta_area)
}
\arguments{
\item{naomi_data}{Naomi object of class "naomi_data" and "naomi_mf"}

\item{indicators}{Naomi indicators created by \code{extract_indicators()}}

\item{meta_areas}{}
}
\description{
Align model data inputs and model estimates
}
