% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-programme.R
\name{anc_prepare_hintr}
\alias{anc_prepare_hintr}
\title{Prepare ANC input data for web interface}
\usage{
anc_prepare_hintr(data)
}
\arguments{
\item{data}{Data frame of input ANC data}
}
\value{
Data frame with additional columns
}
\description{
This function adds additional columns required for plotting ANC data
in the Naomi web app. It will
\itemize{
\item Calculate prevalence and ART coverage
\item Add sample size
}
}
