% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warnings.R
\name{anc_spectrum_warning}
\alias{anc_spectrum_warning}
\title{Compare aggregated district ANC inputs + spectrum totals}
\usage{
anc_spectrum_warning(anc, shape, pjnz)
}
\arguments{
\item{anc}{Path to file containing ANC data or ANC data object}

\item{shape}{Path to file containing geojson areas data or areas data object}

\item{pjnz}{Path to zip file containing spectrum pjnz file/s}
}
\description{
Generate warning if aggregated district art totals do not match spectrum totals
}
\keyword{internal}
