% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{anc_testing_prev_mf}
\alias{anc_testing_prev_mf}
\alias{anc_testing_artcov_mf}
\alias{artnum_mf}
\title{Prepare Model Frames for Programme Datasets}
\usage{
anc_testing_prev_mf(quarter_ids, anc_testing, naomi_mf)

anc_testing_artcov_mf(quarter_ids, anc_testing, naomi_mf)

artnum_mf(quarter_id, art_number, naomi_mf)
}
\arguments{
\item{quarter_ids}{Quarter IDs (possibly multiple)}

\item{anc_testing}{ART data frame}

\item{naomi_mf}{Naomi model frame}

\item{quarter_id}{Quarter ID (single quarter)}

\item{art_number}{Number on ART}
}
\value{
Calculated prevalence
}
\description{
Prepare Model Frames for Programme Datasets
}
