% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-art.R
\name{apply_art_adjustment}
\alias{apply_art_adjustment}
\title{Adjusted reported number on ART based on national/regional adjustment factor in
Spectrum and/or subnational adjustment factor specified ART data}
\usage{
apply_art_adjustment(art, shape, art_spectrum_comparison)
}
\arguments{
\item{art}{Subnational ART data read by \code{read_art_number()}}

\item{shape}{Shapefile read by \code{read_area_merged()}}

\item{art_spectrum_comparison}{Created by \code{prepare_art_spectrum_comparison()}}
}
\description{
Adjusted reported number on ART based on national/regional adjustment factor in
Spectrum and/or subnational adjustment factor specified ART data
}
\keyword{internal}
