% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-programme.R
\name{art_prepare_hintr}
\alias{art_prepare_hintr}
\title{Prepare ART input data for web interface}
\usage{
art_prepare_hintr(data)
}
\arguments{
\item{data}{Data frame of input ART data}
}
\value{
Data frame with additional columns
}
\description{
This function adds additional columns required for plotting ART data
in the Naomi web app. It will
\itemize{
\item Add sample size
}
}
