% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warnings.R
\name{art_spectrum_warning}
\alias{art_spectrum_warning}
\title{Compare aggregated district ART inputs + spectrum totals}
\usage{
art_spectrum_warning(art, shape, pjnz)
}
\arguments{
\item{art}{Path to file containing ART data or ART data object}

\item{shape}{Path to file containing geojson areas data or areas data object}

\item{pjnz}{Path to zip file containing spectrum pjnz file/s}
}
\description{
Generate warning if aggregated district art totals do not match spectrum totals
}
\keyword{internal}
