% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{artnum_mf}
\alias{artnum_mf}
\title{Build artnum model frame}
\usage{
artnum_mf(calendar_quarter, art_number, naomi_mf)
}
\arguments{
\item{calendar_quarter}{Calendar quarter}
}
\description{
Build artnum model frame
}
\details{
Number on ART at desired quarter are linearly interpolated within the dataset.
If the desired quarter is before the earliest data, the first value may
be carried back by up to one year (four quarters). Data are never carried forward
}
