% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs-spectrum.R
\name{assert_pjnz_shiny90}
\alias{assert_pjnz_shiny90}
\title{Check whether PJNZ contains .shiny90 file}
\usage{
assert_pjnz_shiny90(pjnz)
}
\arguments{
\item{pjnz}{file path to PJNZ}
}
\value{
Logical whether PJNZ file contains a .shiny90 file
}
\description{
Check whether PJNZ contains .shiny90 file
}
