% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report-plots.R
\name{bar_plotly}
\alias{bar_plotly}
\title{Plotly barplot comparing geographical distribution of data inputs and naomi outputs}
\usage{
bar_plotly(df, ind, quarter, age_disag = "Y015_049", sex_disag = "both")
}
\arguments{
\item{df}{Inputs_outputs dataframe containing matched model estimates and data inputs.}

\item{ind}{Indicator filter.}

\item{quarter}{Calendar quarter filter.}

\item{age_disag}{Age group filter.}

\item{sex_disag}{Sex filter.}
}
\description{
Plotly barplot comparing geographical distribution of data inputs and naomi outputs
}
