% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-time-series.R
\name{build_hierarchy_label}
\alias{build_hierarchy_label}
\title{Return the translated label & description for a set of plot types}
\usage{
build_hierarchy_label(meta_areas)
}
\arguments{
\item{meta_areas}{dataframe containing}
}
\value{
For each plot type the label and description as a list of lists
containing id, label and description
}
\description{
Return the translated label & description for a set of plot types
}
