% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmb-model-r-implementation.R
\name{bym2_conditional_lpdf}
\alias{bym2_conditional_lpdf}
\title{Log posterior density of BYM2}
\usage{
bym2_conditional_lpdf(x, u, sigma, phi, Q)
}
\arguments{
\item{x}{vector of random effects.}

\item{u}{vector of spatial component of random effect.}

\item{sigma}{marginal standard deviation (>0).}

\item{phi}{proportion of marginal variance explained by spatial structured
component u (phi \in \link{0, 1}).}

\item{Q}{scaled structure matrix for spatial component.}
}
\value{
Log probability density of x and u.
}
\description{
Log posterior density of BYM2
}
\details{
Log posterior density of BYM2 with INLA conditional parameterisation

Calculate the joint LPDF of parameter vector (x, u) where
x = sigma * (sqrt(phi) * u + sqrt(1-phi) * v) with u a ICAR structured
component u ~ N(0, Q^-1) and v is an IID effect v ~ N(0, 1). Calculation
proceeds by conditioning P(<x, u>) = P(x | u) * P(u). See Reibler et al.
Section 3.4.

The $\sqrt(2\pi)^{-2*n}$ and $|Q|^{1/2}$ terms are dropped.
Returns the \emph{positive} log PDF (different from builtin TMB
functions. Thus should typically be implemented as \verb{nll -= bym2_conditional_lpdf(...)}.
}
