% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{cache_invariant}
\alias{cache_invariant}
\title{Make a function cache invariant}
\usage{
cache_invariant(name, fn)
}
\arguments{
\item{name}{The name of the function in the cache}

\item{fn}{The function itself}
}
\value{
Cache invariant version of fn
}
\description{
This first time a function from this is called it will run the function
itself and save the output in memory. On subsequent calls the in
memory cached version will be returned. This cache is separate by
language.
}
\keyword{internal}
