% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-survey.R
\name{calc_survey_hiv_indicators}
\alias{calc_survey_hiv_indicators}
\title{Calculate age/sex/area stratified survey estimates for biomarker outcomes}
\usage{
calc_survey_hiv_indicators(survey_meta, survey_regions, survey_clusters,
  survey_individuals, survey_biomarker, areas, sex = c("male", "female",
  "both"), age_group_id = NULL, area_top_level = min(areas$area_level),
  area_bottom_level = max(areas$area_level))
}
\description{
Calculate age/sex/area stratified survey estimates for biomarker outcomes
}
\details{
All other data will be subsetted based on the \code{survey_id} values appearing in
survey_meta, so if only want to calculate for a subset of surveys it is
sufficient to pass subset for survey_meta and full data frames for the others.

Much of this function needs to be parsed out into more generic functions and
rewritten to be more efficient.
\itemize{
\item Age group would be more efficient if traversing a tree structure.
\item Need generic function to calculate
\item Flexibility about age/sex stratifications to calculate.
}
}
