% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate.R
\name{calculate_logit_adjustment}
\alias{calculate_logit_adjustment}
\alias{calibrate_logistic_adjust_numerator}
\alias{calibrate_logistic_one}
\alias{calibrate_proportional_one}
\alias{calibrate_one}
\title{Helper functions for implementing calibration}
\usage{
calculate_logit_adjustment(proportion_raw, denominator_new, target_val)

calibrate_logistic_adjust_numerator(theta, l, d)

calibrate_logistic_one(proportion_raw, denominator_new, target_val)

calibrate_proportional_one(value_raw, target_val)

calibrate_one(
  value_raw,
  proportion_raw,
  denominator_new,
  target_val,
  calibrate_method
)
}
\arguments{
\item{proportion_raw}{vector of initial input proportions}

\item{denominator_new}{vector of denomin}

\item{target_val}{single value for the adjusted total}
}
\description{
Helper functions for implementing calibration
}
\details{
These functions calibrate input proportions so that
\code{sum(proportion_adj * denominator_new) == target_val},
where \code{proportion_adj} is adjusted results of input
\code{proportion_raw} adjusted by a fixed odds ratio.

For function interface convenience, the parameter \code{target_val}
may be a single number or a vector with the same value
repeated (typically of same length as \code{proportion_raw} and
\code{denominator_new}), but the value must be identical and only
the first value of the vector will be used.

\code{calculate_logit_adjustment()} returns the calculated logit
adjustment value.

\code{calibrate_logistic_one()} returns the adjusted vector (same
length as \code{proportion_raw}.
}
\keyword{internal}
