% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-anc.R
\name{calculate_prevalence_art_coverage}
\alias{calculate_prevalence_art_coverage}
\title{Calculate prevalence and ART coverage from ANC input data}
\usage{
calculate_prevalence_art_coverage(data)
}
\arguments{
\item{data}{Data frame of input ANC data}
}
\value{
Data frame with prevalence and ART coverage
}
\description{
Calculate prevalence and ART coverage from ANC input data
}
