% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputs.R
\name{calibrate_outputs}
\alias{calibrate_outputs}
\title{Calibrate naomi model outputs}
\usage{
calibrate_outputs(
  output,
  naomi_mf,
  spectrum_plhiv_calibration_level,
  spectrum_plhiv_calibration_strat,
  spectrum_artnum_calibration_level,
  spectrum_artnum_calibration_strat,
  spectrum_infections_calibration_level,
  spectrum_infections_calibration_strat
)
}
\description{
Calibrate naomi model outputs
}
