% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-areas.R
\name{check_boundaries}
\alias{check_boundaries}
\title{Check full and aggregated}
\usage{
check_boundaries(sh1, sh2 = NULL)
}
\arguments{
\item{sh1}{Bottom shapefile with red boundaries}

\item{sh2}{Top shapefile with red boundaries}
}
\description{
This function is useful for checking level of coarseness
of a simplified versus raw shapefile and any slivers
in a shapefile.
}
