% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-areas.R
\name{compare_boundaries}
\alias{compare_boundaries}
\title{Compare boundaries of two shapefiles by overlaying them}
\usage{
compare_boundaries(sh1, sh2 = NULL, aggregate = FALSE)
}
\arguments{
\item{sh1}{is bottom shapefile with red boundaries}

\item{sh2}{is top shapefile with red boundaries}

\item{aggregate}{whether to aggregate shapefiles}
}
\description{
Compare boundaries of two shapefiles by overlaying them
}
