% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/car.R
\name{create_adj_matrix}
\alias{create_adj_matrix}
\title{Create adjacency matrix from spatial polygons}
\usage{
create_adj_matrix(sh)
}
\arguments{
\item{sh}{a spatial polygons object}
}
\value{
an adjacency matrix with 1s and 0s
}
\description{
Create adjacency matrix from spatial polygons
}
\details{
Input \code{sh} can be either sf class or SpatialPolygons from \code{sp}. Uses \code{spdep}
package to convert shapefile to neighbor list to adjacency matrix.
}
