% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs-spectrum.R
\name{cut_naomi_age_group}
\alias{cut_naomi_age_group}
\title{Cut Five Year Age Groups}
\usage{
cut_naomi_age_group(age)
}
\arguments{
\item{age}{a vector of ages.}
}
\value{
a vector of strings with five year age groups.
}
\description{
Wrapper for \verb{[cut()]} to return five year age groups with
}
\seealso{
get_age_groups
}
