% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-model.R
\name{data_type_labels}
\alias{data_type_labels}
\title{Get id to label mapping for calibration plot data type}
\usage{
data_type_labels()
}
\value{
List of ID & label mappings
}
\description{
Get id to label mapping for calibration plot data type
}
