% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{demo_survey_biomarker}
\alias{demo_survey_biomarker}
\title{Malawi survye biomarkers}
\format{
Data frame
}
\source{
Malawi
}
\usage{
demo_survey_biomarker
}
\description{
Malawi survye biomarkers
}
\keyword{datasets}
