% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{demo_survey_clusters}
\alias{demo_survey_clusters}
\title{Malawi survey clusters}
\format{
Data frame
}
\source{
Malawi
}
\usage{
demo_survey_clusters
}
\description{
Malawi survey clusters
}
\keyword{datasets}
