% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{demo_survey_hiv_indicators}
\alias{demo_survey_hiv_indicators}
\title{Malawi HIV indicators}
\format{
Data frame
}
\source{
Malawi
}
\usage{
demo_survey_hiv_indicators
}
\description{
Malawi HIV indicators
}
\keyword{datasets}
