% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{demo_survey_regions}
\alias{demo_survey_regions}
\title{Malawi survey regions}
\format{
Data frame
}
\source{
Malawi
}
\usage{
demo_survey_regions
}
\description{
Malawi survey regions
}
\keyword{datasets}
