% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputs.R
\name{disaggregate_0to4_outputs}
\alias{disaggregate_0to4_outputs}
\title{Disaggregate age 0-4 outputs to <1 / 1-4}
\usage{
disaggregate_0to4_outputs(output, naomi_mf)
}
\arguments{
\item{output}{a \code{naomi_output} object}

\item{naomi_mf}{a \code{naomi_mf} object}
}
\description{
Disaggregate output indicators for the 0-4 age group proportional to
Spectrum distribution for each age group.
}
