% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputs.R
\name{display_number}
\alias{display_number}
\title{Format numerical values for display}
\usage{
display_number(x, accuracy = 1e-04, scale = 1, prefix = "", suffix = "")
}
\arguments{
\item{x}{numerical vector.}

\item{accuracy}{A number to round to. E.g. use '0.01' to show 2
decimal places. Accuracy is applied after scaling.}

\item{scale}{A scaling factor: 'x' is multipled by 'scale' before
any formatting.}

\item{prefix}{Symbols to display before the value.}

\item{suffix}{Symbols to display after the value.}
}
\value{
A character vector of \code{length(x)}.
}
\description{
Format numerical values for display
}
\details{
This is modelled after \code{\link[scales:number]{scales::number()}}.

All arguments must be either length 1 or length(x).

'NA' values in \code{prefix} or \code{suffix} are replaced by "".
}
\examples{

x <- c(50.97617, 767.07075, 188.78788, 3.49416, 718.98248)
accuracy <- c(0.01, 1, 100, 0.1, 0.001)
scale <- c(1, 1, 1, 100, 1)
prefix <- c("$", "", "", "", "")
suffix <- c("", "", "", "\%", "")

display_number(x, accuracy, scale, prefix, suffix)
paste0(prefix, round(scale * x / accuracy) * accuracy, suffix)

}
