% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report-plots.R
\name{district_barplot}
\alias{district_barplot}
\title{Create district bar plot}
\usage{
district_barplot(
  naomi_output,
  indicator,
  age,
  sex_disag,
  level = NULL,
  calendar_quarter = 2,
  label_format,
  x_title = NULL,
  fig_title = NULL,
  legend_title = NULL,
  colour_palette = "blue",
  above_label = t_("ABOVE"),
  below_label = t_("BELOW")
)
}
\arguments{
\item{naomi_output}{Naomi output object or path to naomi output zip folder}

\item{calendar_quarter}{Naomi projection quarter to filter data to, default is calendar_quarter_t2}

\item{x_title}{Title for the figure x axis, blank by default}

\item{fig_title}{Title for the figure, blank by default,}

\item{colour_palette}{Names or hexcode value for colours to differentiate districts with values above and below the national average. Can be specified as a preset colour palette using "red", "green" and "blue" or manually as a named list: cols <- c(above_colour = "red", below_colour = "blue").}

\item{above_label}{Label for districts with values higher than national average, default is "Above"}

\item{below_label}{Label for districts with values higher than national average, default is "Below"}

\item{y_title}{Title for the figure y axis, default is "Age Group"}
}
\value{

}
\description{
Create district bar plot
}
