% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report-plots.R
\name{dropdown_buttons}
\alias{dropdown_buttons}
\title{Automatically make drop drop buttons for plotly filter}
\usage{
dropdown_buttons(df, var)
}
\arguments{
\item{df}{Input data frame}

\item{var}{Column}
}
\value{
Drop down buttons for plotly filter
}
\description{
Automatically make drop drop buttons for plotly filter
}
