% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-survey.R
\name{expand_survey_clusters}
\alias{expand_survey_clusters}
\title{Expand list of clusters at each area level}
\usage{
expand_survey_clusters(survey_clusters, survey_regions, areas,
  top_level = min(areas$area_level),
  bottom_level = max(areas$area_level))
}
\description{
This function recursively expands the list of clusters to produce a list
of survey clusters within areas at each level.
}
\details{
TODO: These should be examples - where is areas_long.rds now?
areas_long <- readRDS(here::here("data/areas/areas_long.rds"))
survey_clusters <- readRDS(here::here("data/survey/survey_clusters.rds"))
survey_regions <- readRDS(here::here("data/survey/survey_regions.rds"))

expand_survey_clusters(survey_clusters, areas_long)

Get clusters at level 1 areas only
expand_survey_clusters(survey_clusters, areas_long, top_level = 1, bottom_level = 1)
}
\keyword{Recursion}
\keyword{internal}
