% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs-spectrum.R
\name{extract_pjnz_naomi}
\alias{extract_pjnz_naomi}
\title{Extract Model Inputs from Spectrum PJNZ}
\usage{
extract_pjnz_naomi(pjnz)
}
\arguments{
\item{pjnz}{Filepath to Spectrum PJNZ file.}
}
\value{
A \code{data.frame} with Spectrum indicators.
}
\description{
Extract Model Inputs from Spectrum PJNZ
}
\examples{
pjnz <- system.file("extdata/mwi2019.pjnz", package = "naomi")
spec <- extract_pjnz_naomi(pjnz)

}
