% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs-spectrum.R
\name{extract_pjnz_program_data}
\alias{extract_pjnz_program_data}
\title{Extract ART and ANC testing program data inputs from Spectrum PJNZ}
\usage{
extract_pjnz_program_data(pjnz_list)
}
\arguments{
\item{pjnz_list}{Vector of filepaths to Spectrum PJNZ file.}
}
\value{
A list with a two \code{data.frame}s of ANC testing data and
number on ART, respectively.
}
\description{
Extract ART and ANC testing program data inputs from Spectrum PJNZ
}
\examples{
pjnz <- system.file("extdata/demo_mwi2019.PJNZ", package = "naomi")
spec <- extract_pjnz_program_data(pjnz)

}
