% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs-shiny90.R
\name{extract_shiny90_age_sex}
\alias{extract_shiny90_age_sex}
\title{Export proportion aware by five year age group from Shiny90}
\usage{
extract_shiny90_age_sex(shiny90_path, years = NULL)
}
\arguments{
\item{shiny90_path}{file path to .shiny90 digest file.}

\item{years}{year(s) to generate estimates; an integer or a vector of integers.
If NULL, all years available in estimates are returned (default).}
}
\value{
A data frame consisting of the number of PLHIV, aware of status
and on ART by sex and five year age group 15-49 and age 50+ for
specified \code{years}.
}
\description{
Export estimates for proportion aware of status from a Shiny90
output file to five-year age groups 15-19 to 50+.
}
\details{
The 'artnum' divided by 'plhiv' columns in the output will give
a different ART coverage than Spectrum output for the same year
because these outputs are based on the internal mid-year ART
coverage in Spectrum, while Spectrum reports an end of year (Dec 31)
ART coverage.
}
\examples{

pjnz <- system.file("extdata/demo_mwi2019.PJNZ", package = "naomi")
shiny90dir <- tempfile()
utils::unzip(pjnz, "malawi.zip.shiny90", exdir = shiny90dir)
shiny90_path <- file.path(shiny90dir, "malawi.zip.shiny90")

extract_shiny90_age_sex(shiny90_path, year = 2010:2019)

}
