% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmb-model.R
\name{fit_tmb}
\alias{fit_tmb}
\title{Fit TMB model}
\usage{
fit_tmb(tmb_input, outer_verbose = TRUE, inner_verbose = FALSE)
}
\arguments{
\item{tmb_input}{Model input data}

\item{outer_verbose}{If TRUE print function and parameters every iteration}

\item{inner_verbose}{If TRUE then disable tracing information from TMB}
}
\value{
Fit model.
}
\description{
Fit TMB model
}
