% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warnings.R
\name{format_spectrum_total_warning}
\alias{format_spectrum_total_warning}
\title{Format spectrum total warnings}
\usage{
format_spectrum_total_warning(data_merged, key, location, age_disag = FALSE)
}
\arguments{
\item{data_merged}{Aggregated program data merged with spectrum totals}

\item{key}{Translation key for warnings}

\item{location}{Location where warning should be displayed}

\item{age_disag}{Logical if age should be included in warning labels. Default is FALSE.}
}
\description{
Format warnings with translated key to be displayed when spectrum totals
do not match aggregated district level programme data
}
\keyword{internal}
