% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-areas.R
\name{gather_areas}
\alias{gather_areas}
\title{Convert nested hierarchy from wide to long format}
\usage{
gather_areas(x)
}
\arguments{
\item{x}{Wide format nested hierarchy.}
}
\description{
Convert nested hierarchy from wide to long format
}
