% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputs.R
\name{generate_output_summary_report}
\alias{generate_output_summary_report}
\title{Generate and save summary report at specified path}
\usage{
generate_output_summary_report(report_path, output_zip, quiet = FALSE)
}
\arguments{
\item{report_path}{Path to save summary report at}

\item{output_zip}{Path to model outputs zip file}

\item{quiet}{Suppress printing of the pandoc command line}
}
\value{
Path to summary report
}
\description{
Generate and save summary report at specified path
}
\keyword{internal}
