% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{get_age_group_id_out}
\alias{get_age_group_id_out}
\title{Get age group ids for output}
\usage{
get_age_group_id_out(age_group_ids)
}
\arguments{
\item{age_group_ids}{Age group ids.}
}
\description{
Ensures that age groups are fully spanned by modelled
ages.
}
\keyword{internal}
