% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{get_age_group_out}
\alias{get_age_group_out}
\title{Get age group ids for output}
\usage{
get_age_group_out(age_groups)
}
\arguments{
\item{age_groups}{Modelled age groups. Assumed to
be non-overlapping.}
}
\description{
Ensures that age groups are fully spanned by modelled
ages.
}
\keyword{internal}
