% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-time-series.R
\name{get_anc_metadata}
\alias{get_anc_metadata}
\title{Metadata for ANC plot types used in the input time series
plots. Mainly for internal use.}
\usage{
get_anc_metadata(anc)
}
\arguments{
\item{anc}{Path to file containing ANC data or ANC data object}
}
\value{
List of \code{plot_types} and \code{cols_keep} (columns that are
present in the actual data out of the known list of columns)
}
\description{
Metadata for ANC plot types used in the input time series
plots. Mainly for internal use.
}
