% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-time-series.R
\name{get_art_metadata}
\alias{get_art_metadata}
\title{Metadata for ART plot types used in the input time series
plots. Mainly for internal use.}
\usage{
get_art_metadata(art)
}
\arguments{
\item{anc}{Path to file containing ART data or ART data object}
}
\value{
List of \code{plot_types} and \code{cols_keep} (columns that are
present in the actual data out of the known list of columns)
and \code{calendar_quarters}
}
\description{
Metadata for ART plot types used in the input time series
plots. Mainly for internal use.
}
