% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-options.R
\name{get_calibration_option_labels}
\alias{get_calibration_option_labels}
\title{Map calibration option ID to JSON calibration option labels}
\usage{
get_calibration_option_labels(options)
}
\arguments{
\item{options}{Key-value (calibration option name - calibration option ID)
list of model options to be mapped.}
}
\value{
Mapped key-value (calibration option name - calibration option label)
list of model options
}
\description{
Map calibration option ID to JSON calibration option labels
}
