% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{get_colour_scale}
\alias{get_colour_scale}
\title{Get colour scale and ranges for a particular country}
\usage{
get_colour_scale(country = "default")
}
\arguments{
\item{country}{Country to get scale for or return default scale if left
blank.}
}
\value{
List of scale information including colour as a d3 scale chromatic
function name, whether to invert the scale and a min and max value for the
scale.
}
\description{
Get colour scale and ranges for a particular country
}
\keyword{internal}
