% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-options.R
\name{get_country_option_defaults}
\alias{get_country_option_defaults}
\title{Get option defaults for a specific country}
\usage{
get_country_option_defaults()
}
\arguments{
\item{iso3}{iso3 code for the country}
}
\value{
The option defaults for country, if no defaults found then
}
\description{
Get option defaults for a specific country
}
