% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{get_five_year_age_groups}
\alias{get_five_year_age_groups}
\title{Get 5 year age groups}
\usage{
get_five_year_age_groups()
}
\value{
ID for 5 year age groups
}
\description{
Get 5 year age groups
}
\examples{
get_five_year_age_groups()
}
