% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputs.R
\name{get_meta_indicator}
\alias{get_meta_indicator}
\title{Get indicator metadata}
\usage{
get_meta_indicator()
}
\value{
data.frame of indicator ids, labels, descriptions, and parameter mapping.
}
\description{
Get indicator metadata
}
\examples{
get_meta_indicator()
}
