% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{get_metadata}
\alias{get_metadata}
\title{Get indicator metadata as a data frame}
\usage{
get_metadata()
}
\value{
Metadata about indicators as a list
}
\description{
Get indicator metadata as a data frame
}
\examples{
get_metadata()
}
