% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-survey.R
\name{get_mid_calendar_quarter}
\alias{get_mid_calendar_quarter}
\title{Find Calendar Quarter Midpoint of Two Dates}
\usage{
get_mid_calendar_quarter(start_date, end_date)
}
\arguments{
\item{start_date}{vector coercibel to Date}

\item{end_date}{vector coercibel to Date}
}
\value{
A vector of calendar quarters
}
\description{
Find Calendar Quarter Midpoint of Two Dates
}
\examples{
start <- c("2005-04-01", "2010-12-01", "2016-01-01")
end <-c("2005-08-01", "2011-05-01", "2016-06-01")

mid_calendar_quarter <- get_mid_calendar_quarter(start, end)

}
