% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-options.R
\name{get_model_calibration_options}
\alias{get_model_calibration_options}
\title{Get JSON representing model calibration options}
\usage{
get_model_calibration_options()
}
\value{
Model calibration options.
}
\description{
Get JSON representing model calibration options
}
\examples{
get_model_calibration_options()
}
