% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-options.R
\name{get_model_run_options}
\alias{get_model_run_options}
\title{Get JSON representing model run options}
\usage{
get_model_run_options()
}
\value{
Validated model run options JSON.
}
\description{
This reads JSON file of model run options which describe how the Naomi
front end should display input options. It validates these against a schema
to ensure it can create a UI.
}
\examples{
get_model_run_options()
}
