% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-time-series.R
\name{get_plot_type_column_metadata}
\alias{get_plot_type_column_metadata}
\title{Return the translated label & description for a set of plot types}
\usage{
get_plot_type_column_metadata(plot_type)
}
\arguments{
\item{plot_type}{Plot type ids}
}
\value{
For each plot type the label and description as a list of lists
containing id, label and description
}
\description{
Return the translated label & description for a set of plot types
}
