% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs-spectrum.R
\name{get_spec_aggr_interpolation}
\alias{get_spec_aggr_interpolation}
\title{Interpolate Spectrum to quarter_id}
\usage{
get_spec_aggr_interpolation(spec_aggr, calendar_quarter_out)
}
\description{
Interpolate Spectrum to quarter_id
}
