% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-model.R
\name{hintr_calibrate}
\alias{hintr_calibrate}
\title{Calibrate hintr_output}
\usage{
hintr_calibrate(output, calibration_options)
}
\arguments{
\item{output}{A hintr_output object}

\item{calibration_options}{A set of calibration options}
}
\value{
Calibrated hintr_output object
}
\description{
Take a previously generated hintr_output object and calibrate. Format
response as another hintr_output object.
}
