% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hintr-plot-data.R
\name{hintr_comparison_plot}
\alias{hintr_comparison_plot}
\title{Get data for hintr comparison plot}
\usage{
hintr_comparison_plot(output)
}
\arguments{
\item{output}{A hintr_output object}
}
\value{
Calibrated, unadjusted and spectrum estimates of indicators
}
\description{
Takes hintr_output object, reads data and prepares data in format needed
for plotting comparison barchart to compare input and output data.
}
