% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloads.R
\name{hintr_prepare_coarse_age_group_download}
\alias{hintr_prepare_coarse_age_group_download}
\title{Prepare coarse age group download}
\usage{
hintr_prepare_coarse_age_group_download(
  output,
  path = tempfile(fileext = ".zip")
)
}
\arguments{
\item{output}{hintr output object}

\item{path}{Path to save output file}
}
\value{
Path to output file and metadata for file
}
\description{
Prepare coarse age group download
}
