% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloads.R
\name{hintr_prepare_shipp_download}
\alias{hintr_prepare_shipp_download}
\title{Prepare SHIPP tool download}
\usage{
hintr_prepare_shipp_download(output, pjnz, path = tempfile(fileext = ".xlsx"))
}
\arguments{
\item{pjnz}{Path to input PJNZ file}

\item{path}{Path to save output file}

\item{hintr_output}{object}
}
\value{
Path to output file and metadata for file
}
\description{
Prepare SHIPP tool download
}
