% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloads.R
\name{hintr_prepare_summary_report_download}
\alias{hintr_prepare_summary_report_download}
\title{Prepare summary report download}
\usage{
hintr_prepare_summary_report_download(
  output,
  path = tempfile(fileext = ".html")
)
}
\arguments{
\item{path}{Path to save output file}

\item{hintr_output}{object}
}
\value{
Path to output file and metadata for file
}
\description{
Prepare summary report download
}
