% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-model.R
\name{hintr_run_model}
\alias{hintr_run_model}
\title{Run the model and save output to file}
\usage{
hintr_run_model(data, options, output_path = tempfile(),
  spectrum_path = tempfile(fileext = ".zip"),
  summary_path = tempfile(fileext = ".zip"))
}
\arguments{
\item{data}{List of paths to input data files.}

\item{options}{List of model run options (see details).}

\item{output_path}{Path to store output indicators as an RDS at.}

\item{spectrum_path}{Path to store spectrum digest file at.}

\item{summary_path}{Path to store summary download zip file at.}
}
\value{
Paths to 3 output files.
}
\description{
This prepares the model inputs from data and options and saves output as
an rds, the spectrum digst and indicators at specified paths.
}
\details{
The \code{data} argument must be a list specifying paths to the following:
\itemize{
\item \code{pjnz}
\item \code{shape}
\item \code{population}
\item \code{survey data}
\item \code{anc} (optional)
\item \code{art} (optional)
}

The \code{options} argument must be a list specifying minimally:
\itemize{
\item area_scope
\item area_level
\item calendar_quarter_t1
\item calendar_quarter_t2
\item survey_prevalence
\item survey_art_coverage
\item survey_recently_infected
\item include_art_t1
\item include_art_t2
\item anc_prevalence_year1
\item anc_prevalence_year2
\item anc_art_coverage_year1
\item anc_art_coverage_year2
}
}
