% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warnings.R
\name{hintr_validate_anc}
\alias{hintr_validate_anc}
\title{Run validation for ANC input}
\usage{
hintr_validate_anc(anc, shape, pjnz)
}
\arguments{
\item{anc}{Path to file containing ANC data or ANC data object}

\item{shape}{Path to file containing geojson areas data or areas data object}

\item{pjnz}{Path to zip file containing spectrum pjnz file/s}
}
\description{
This can throw validation errors or warnings which will be shown to user
in naomi web app
}
