% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warnings.R
\name{hintr_validate_art}
\alias{hintr_validate_art}
\title{Run validation for ART input}
\usage{
hintr_validate_art(art, shape, pjnz)
}
\arguments{
\item{art}{Path to file containing ART data or ART data object}

\item{shape}{Path to file containing geojson areas data or areas data object}

\item{pjnz}{Path to zip file containing spectrum pjnz file/s}
}
\description{
This can throw validation errors or warnings which will be shown to user
in naomi web app
}
