% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warnings.R
\name{hintr_validate_programme_data}
\alias{hintr_validate_programme_data}
\title{Run validation for subnational programme data input}
\usage{
hintr_validate_programme_data(naomi_spectrum_comparison)
}
\arguments{
\item{naomi_spectrum_comparison}{Comparison table of aggregated subnational
Naomi and national Spectrum programme data created by
prepare_art_spectrum_comparison() or prepare_anc_spectrum_comparison()}
}
\description{
This can throw validation errors or warnings which will be shown to user
in naomi web app
}
