% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-population.R
\name{interpolate_population_agesex}
\alias{interpolate_population_agesex}
\title{Log-linear interpolation of age/sex stratified population}
\usage{
interpolate_population_agesex(population_agesex, times = seq(2010.5,
  2019.5, 0.25))
}
\arguments{
\item{population_agesex}{a subset of the population_agesex.}

\item{times}{vector of times to return interpolation.}
}
\value{
A data.frame with same columns as pop_agesex interpolated to \code{times}.
}
\description{
Log-linear interpolation of age/sex stratified population
}
\details{
\code{zoo::na.approx()} is used to interpolate log(population).
}
\examples{
## Interpolate Malawi population at level 2 (Zone) at two time points
population_agesex <- read.csv(
  system.file("extdata/population/population_agesex.csv", package = "naomi"),
  stringsAsFactors = FALSE)
pop_interp <- interpolate_population_agesex(population_agesex, c(2016.25, 2019.75))

}
