% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmb-model-r-implementation.R
\name{ldbinom}
\alias{ldbinom}
\title{Binomial distribution log-density permitting non-integer counts}
\usage{
ldbinom(x, size, prob)
}
\description{
Binomial distribution log-density permitting non-integer counts
}
\examples{

ldbinom(c(1.2, 3, 1.9), c(7.5, 12.4, 4), c(0.3, 0.2, 0.7))

## For integer counts, returns same as dbinom(..., log = TRUE)
x <- c(1, 3, 2)
size <- c(7, 12, 4)
prob <- c(0.3, 0.2, 0.7)
ldbinom(x, size, prob)
dbinom(x, size, prob, log = TRUE)

}
