% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-population.R
\name{log_linear_interp}
\alias{log_linear_interp}
\title{Log-linear interpolation of NA values}
\usage{
log_linear_interp(y, x)
}
\description{
Log-linear interpolation of NA values
}
\examples{
log_linear_interp(c(100, 105, NA, 110), 1:4)

}
