% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shipp.R
\name{logit_scale_prev}
\alias{logit_scale_prev}
\title{Calculate prevalence and PLHIV using logit-scale disaggregation}
\usage{
logit_scale_prev(lor, N_fine, plhiv)
}
\arguments{
\item{lor}{Log odds-ratios}

\item{N_fine}{Number of individuals in each group}

\item{plhiv}{Total number of people living with HIV}
}
\description{
Calculate prevalence and PLHIV using logit-scale disaggregation
}
\keyword{internal}
