% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-helpers.R
\name{make_agyw_testfiles}
\alias{make_agyw_testfiles}
\title{Build JSON from template and a set of params}
\usage{
make_agyw_testfiles(naomi_output)
}
\arguments{
\item{naomi_output}{Calibrated naomi output}
}
\value{
Calibrated naomi output matched to MWI test data on
\code{naomi.resources} to be used to generate the agyw tool.
}
\description{
Build JSON from template and a set of params
}
\keyword{internal}
