% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report-plots.R
\name{map_outputs}
\alias{map_outputs}
\title{Create output map plot}
\usage{
map_outputs(
  geom_data,
  indicator,
  age,
  sex_disag,
  start_colour_scale,
  end_colour_scale,
  fig_title = NULL,
  legend_title = NULL,
  legend_label = ggplot2::waiver(),
  breaks = 6
)
}
\arguments{
\item{geom_data}{Data frame containing indicators output and geometry
filtered to a single quarter}

\item{indicator}{The indicator to generate plot for}

\item{age}{The age group to include in plots}

\item{sex_disag}{Sex diaggregation to use in plot,}

\item{start_colour_scale}{The colour value to start scale at, applied to
lowest value in the data.}

\item{end_colour_scale}{The colour value to end scale at, applied to highest
value in the supplied data,}

\item{fig_title}{Title for the figure, blank by default}

\item{legend_title}{Title for the legend, blank by default}

\item{legend_label}{Legend label}

\item{breaks}{Number of break points to create in scale}
}
\value{

}
\description{
Create output map plot
}
