% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{mwi_anc_testing}
\alias{mwi_anc_testing}
\alias{mwi_art_number}
\title{Programme data: Malawi routine HIV service delivery data}
\format{
TODO: Link to mwi_areas dataset
\code{mwi_art_number} is a dataset with 2042 observations on the following 5 variables:
\describe{
\item{\code{area_id}}{District area id -- see mwi_areas dataset.}
\item{\code{sex}}{Sex as a character vector -- all values are "both".}
\item{\code{age_group_id}}{Value 20 is age 15+, 24 is age 0-14.}
\item{\code{quarter_id}}{Quarter ID (integer).}
\item{\code{current_art}}{Number currently on ART at the end of the quarter.}
}

TODO: Link to mwi_areas dataset
\code{mwi_anc_testing} is a dataset with 923 observations on the following 8 variables:
\describe{
\item{\code{area_id}}{District area id -- see mwi_areas dataset.}
\item{\code{quarter_id}}{Quarter ID (integer).}
\item{\code{anc_clients}}{Number of unique ANC clients during quarter.}
\item{\code{ancrt_known_pos}}{Number of ANC clients already known to be HIV positive prior to first ANC visit and not tested for HIV.}
\item{\code{ancrt_already_art}}{Number of ANC clients already on ART prior to first ANC visit.}
\item{\code{ancrt_tested}}{Number of ANC clients tested for HIV and received result.}
\item{\code{ancrt_test_pos}}{Number of ANC clients testing HIV postive.}
}

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 510 rows and 6 columns.
}
\source{
Department of HIV & AIDS, Minitry of Health, Malawi.
\url{http://hiv.health.gov.mw/index.php/our-documents}
}
\usage{
mwi_anc_testing

mwi_art_number
}
\description{
Datasets containing Malawi National HIV programme quarterly reported data
about:
\itemize{
\item Reported number currently on ART.
\item Number of women receiving antenatal care and final antenatal HIV testing.
}
}
\details{
Data are sourced from quarterly reporting spreadsheets produced via the Quarterly
Integrated HIV Program Reports and publicly available from Malawi Ministry of
Health: Department of HIV & AIDS website.

Data are provided by the Ministry of Health, Malawi with the following disclaimer:
\enumerate{
\item The data in this file are owned by the Ministry of Health (MOH), Malawi.
\item These data are shared in order to support the National HIV Program.
\item Data in this file may not be used by Third Parties for further analysis and
dissemination without prior written approval of the Director of the Department
for HIV and AIDS, MOH.
\item Patient-level data are routinely collected by health facility staff using
standard monitoring tools. This file contains facility-level aggregates that
have been generated by facility staff and that were verified from primary
records during quarterly National HIV Program Supervision (coordinated of the
Department of HIV and AIDS). While every effort is made to ensure high data
quality, individual records may not be complete and accurate.
\item The official interpretation of these data is presented in Quarterly Integrated
HIV Program Reports and any divergent interpretation may be misleading and is
not supported by the MOH. Analysis and interpretation of the data requires
detailed understsanding of the methods and constraints of the Malawi's National
M&E System.
}

The number of ART by age 0-14 and age 15+ is not reported in aggregate quarterly
reporting data. For model illustration purposes, this is approximated as 94\% of
all ART clients are age 15+ based on Spectrum model outputs which were validated
against age distributions from electronic medical records by the Malawi HIV
estimates team.

Dataset could contain one observation per district per quarter for all 32 districts
for all quarters between Q3 2011 and Q3 2018, but the number of observations is
fewer than 32 * 29 = 928 due to a few missing observations.
}
\seealso{
\code{\link{quarter_labels}}, \code{\link{get_age_groups}}.
}
\keyword{datasets}
