% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mwi_area_boundaries}
\alias{mwi_area_boundaries}
\title{Malawi area boundaries.}
\format{
sf object representing area boundaries
}
\source{
Malawi
}
\usage{
mwi_area_boundaries
}
\description{
Malawi area boundaries.
}
\keyword{datasets}
