% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mwi_area_centers}
\alias{mwi_area_centers}
\title{Malawi area centers}
\format{sf object representing area centers}
\source{
Malawi
}
\usage{
mwi_area_centers
}
\description{
Malawi area centers
}
\keyword{datasets}
