% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mwi_area_hierarchy}
\alias{mwi_area_hierarchy}
\title{Malawi area hierarchy}
\format{Data frame}
\source{
Malawi
}
\usage{
mwi_area_hierarchy
}
\description{
Malawi area hierarchy
}
\keyword{datasets}
