% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mwi_area_meta}
\alias{mwi_area_meta}
\title{Malawi area metdata}
\format{Data frame}
\source{
Malawi
}
\usage{
mwi_area_meta
}
\description{
Meta information about area hierarchy invluding admin level names
}
\keyword{datasets}
