% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mwi_population_agesex}
\alias{mwi_population_agesex}
\title{Malawi age-sex stratified population data}
\format{
Data frame
}
\source{
Malawi
}
\usage{
mwi_population_agesex
}
\description{
Malawi age-sex stratified population data
}
\keyword{datasets}
