% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mwi_survey_clusters}
\alias{mwi_survey_clusters}
\title{Malawi survey clusters}
\format{Data frame}
\source{
Malawi
}
\usage{
mwi_survey_clusters
}
\description{
Malawi survey clusters
}
\keyword{datasets}
