% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mwi_survey_individuals}
\alias{mwi_survey_individuals}
\title{Malawi survey individuals}
\format{
Data frame
}
\source{
Malawi
}
\usage{
mwi_survey_individuals
}
\description{
Malawi survey individuals
}
\keyword{datasets}
