% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mwi_survey_meta}
\alias{mwi_survey_meta}
\title{Malawi survey metadata}
\format{
Data frame
}
\source{
Malawi
}
\usage{
mwi_survey_meta
}
\description{
Malawi survey metadata
}
\keyword{datasets}
