% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mwi_survey_regions}
\alias{mwi_survey_regions}
\title{Malawi survey regions}
\format{Data frame}
\source{
Malawi
}
\usage{
mwi_survey_regions
}
\description{
Malawi survey regions
}
\keyword{datasets}
