% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmb-model-r-implementation.R
\name{naomi_objective_function_r}
\alias{naomi_objective_function_r}
\title{R implementation of Naomi model}
\usage{
naomi_objective_function_r(d, p)
}
\arguments{
\item{d}{list of data inputs (from \code{\link[=prepare_tmb_inputs]{prepare_tmb_inputs()}})}

\item{p}{list of parameter values (from \code{\link[=prepare_tmb_inputs]{prepare_tmb_inputs()}})}
}
\value{
A list consisting of (1) reported objects from the TMB model,
and (2) the value of the objective function.
}
\description{
R implementation of Naomi model
}
\details{
This is primarily written to enable stepping through the TMB model
line by line.
}
