% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{naomi_output_frame}
\alias{naomi_output_frame}
\title{Model Frame and Linear Transform for Aggregated Model Outputs}
\usage{
naomi_output_frame(mf_model, area_aggregation)
}
\arguments{
\item{mf_model}{Model frame}

\item{areas}{naomi_areas object.}
}
\description{
Model Frame and Linear Transform for Aggregated Model Outputs
}
