% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traduire.R
\name{naomi_set_language}
\alias{naomi_set_language}
\title{Set naomi language}
\usage{
naomi_set_language(language)
}
\arguments{
\item{language}{Either "en" or "fr"}
}
\description{
Set language to use for naomi messages, etc
}
