% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warnings.R
\name{output_naomi_warning}
\alias{output_naomi_warning}
\title{Generate naomi warning for specific strata of simulated outputs}
\usage{
output_naomi_warning(naomi_output, ind, threshold, locations)
}
\arguments{
\item{ind}{Naomi output indicator}

\item{threshold}{Numerical threshold to trigger warning}

\item{locations}{Location where warning should be displayed in front end}
}
\description{
Generate naomi warning for specific strata of simulated outputs
}
