% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report-plots.R
\name{pop_pyramid_outputs}
\alias{pop_pyramid_outputs}
\title{Create population pyramid plot}
\usage{
pop_pyramid_outputs(
  naomi_output,
  calendar_quarter = 2,
  indicator,
  colour_palette = "blue",
  x_title = NULL,
  y_title = t_("AGE_GROUP"),
  fig_title = NULL,
  legend_label = abs,
  masc_label = t_("SEX_MALE"),
  fem_label = t_("SEX_FEMALE")
)
}
\arguments{
\item{naomi_output}{Naomi output object or path to naomi output zip folder}

\item{calendar_quarter}{Naomi projection quarter to filter data to, default is calendar_quarter_t2}

\item{colour_palette}{Names or hexcode value for right and left side of population pyramid. Can be specified as a preset colour palette using "red", "green" and "blue" or manually as a named list: cols <- c( left_clour = "red", right_colour = "blue").}

\item{x_title}{Title for the figure x axis, blank by default}

\item{y_title}{Title for the figure y axis, default is "Age Group"}

\item{fig_title}{Title for the figure, blank by default,}

\item{masc_label}{Label for male sex group, default is "Male"}

\item{fem_label}{Label for female sex group, default is "Female"}
}
\value{
Pop pyramid
}
\description{
Create population pyramid plot
}
