% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-comparison.R
\name{prepare_anc_spectrum_comparison}
\alias{prepare_anc_spectrum_comparison}
\title{Compare aggregated subnational ART inputs + spectrum totals for comparison table}
\usage{
prepare_anc_spectrum_comparison(anc, shape, pjnz)
}
\arguments{
\item{shape}{Path to file containing geojson areas data or areas data object}

\item{pjnz}{Path to zip file containing spectrum pjnz file/s}

\item{art}{Path to file containing ART data or ART data object}
}
\description{
Compare aggregated subnational ART inputs + spectrum totals for comparison table
}
\keyword{internal}
