% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-comparison.R
\name{prepare_art_spectrum_comparison}
\alias{prepare_art_spectrum_comparison}
\title{Compare aggregated district ART inputs + spectrum totals
Compare aggregated subnational ART inputs + spectrum totals for comparison table}
\usage{
prepare_art_spectrum_comparison(art, shape, pjnz)
}
\arguments{
\item{art}{Path to file containing ART data or ART data object}

\item{shape}{Path to file containing geojson areas data or areas data object}

\item{pjnz}{Path to zip file containing spectrum pjnz file/s}
}
\description{
Compare aggregated district ART inputs + spectrum totals
Compare aggregated subnational ART inputs + spectrum totals for comparison table
}
\keyword{internal}
