% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inla-model.R
\name{prepare_inla_inputs}
\alias{prepare_inla_inputs}
\title{Prepare inputs for INLA}
\usage{
prepare_inla_inputs(naomi_data)
}
\arguments{
\item{naomi_data}{Naomi data object}
}
\value{
Inputs ready for INLA
}
\description{
Prepare inputs for INLA
}
\seealso{
\link{select_naomi_data}
}
