% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-time-series.R
\name{prepare_input_time_series_anc}
\alias{prepare_input_time_series_anc}
\title{Prepare data for ANC input time series plots}
\usage{
prepare_input_time_series_anc(anc, shape)
}
\arguments{
\item{anc}{Path to file containing ANC data}

\item{shape}{Path to file containing geojson areas data}
}
\value{
Data formatted for plotting input time series containing columns
area_id, area_name, area_level, area_level_label, age_group, time_period,
time_step, plot and value
}
\description{
Take uploaded ANC and shape file paths and format as data which
can be used to draw input time series graphs.
}
