% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-time-series.R
\name{prepare_input_time_series_art}
\alias{prepare_input_time_series_art}
\title{Prepare data for ART input time series plots}
\usage{
prepare_input_time_series_art(art, shape)
}
\arguments{
\item{art}{Path to file containing ART data or ART data object}

\item{shape}{Path to file containing geojson areas data or area data object}
}
\value{
Data formatted for plotting input time series containing columns
area_id, area_name, area_level, area_level_label, parent_area_id, area_sort_order,
time_period, year, quarter, calendar_quarter, area_hierarchy, plot, value and
missing_ids
}
\description{
Take uploaded ART and shape file paths and format as data which
can be used to draw input time series graphs.
}
