% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-comparison.R
\name{prepare_spectrum_naomi_comparison}
\alias{prepare_spectrum_naomi_comparison}
\title{Compare aggregated subnational Naomi + spectrum totals for comparison table}
\usage{
prepare_spectrum_naomi_comparison(art, anc, shape, pjnz)
}
\arguments{
\item{art}{Path to file containing ART data or ART data object}

\item{anc}{Path to file containing ART data or ART data object}

\item{shape}{Path to file containing geojson areas data or areas data object}

\item{pjnz}{Path to zip file containing spectrum pjnz file/s}
}
\description{
Compare aggregated subnational Naomi + spectrum totals for comparison table
}
