% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmb-model.R
\name{prepare_tmb_inputs}
\alias{prepare_tmb_inputs}
\title{Prepare inputs for TMB model.}
\usage{
prepare_tmb_inputs(naomi_mf, prev_dat, artcov_dat, recent_dat,
  anc_prev_t1_dat, anc_prev_t2_dat, anc_artcov_t1_dat, anc_artcov_t2_dat,
  artnum_t1_dat, artnum_t2_dat)
}
\arguments{
\item{naomi_mf}{Naomi model frame}

\item{prev_dat}{Prevalence data}

\item{artcov_dat}{ART coverage data}

\item{recent_dat}{Recent infection data}

\item{anc_prev_t1_dat}{ANC prevalence at time 1}

\item{anc_prev_t2_dat}{ANC prevelance at time 2}

\item{anc_artcov_t1_dat}{ANC ART coverage at time 1}

\item{anc_artcov_t2_dat}{ANC ART coverage at time 2}

\item{artnum_t1_dat}{Number on ART at time 1}

\item{artnum_t2_dat}{Number on ART at time 2}
}
\value{
Inputs ready for TMB model
}
\description{
Prepare inputs for TMB model.
}
