% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warnings.R
\name{programme_data_warning}
\alias{programme_data_warning}
\title{Warning for aggregated subnational data input snot equal to spectrum totals}
\usage{
programme_data_warning(naomi_spectrum_comparison)
}
\arguments{
\item{naomi_spectrum_comparison}{Comparison table of aggregated subnational
Naomi and national Spectrum programme data created by
prepare_art_spectrum_comparison() or prepare_anc_spectrum_comparison()}
}
\description{
Generate warning if aggregated subnational totals do not match spectrum totals
}
\keyword{internal}
