% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputs.R
\name{read_hintr_output}
\alias{read_hintr_output}
\title{Read hintr output}
\usage{
read_hintr_output(path)
}
\arguments{
\item{path}{Path to the file}
}
\value{
The read data
}
\description{
Read the hintr model output or plot data saved as a qs or an rds file.
This is the data saved from hintr_run_model or hintr_calibrate
before the output zip is generated. This uses the file extension
to identify the reading function to use.
}
