% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{read_metadata}
\alias{read_metadata}
\title{Get indicator metadata as a data frame}
\usage{
read_metadata()
}
\value{
Metadata about indicators as a list
}
\description{
Get indicator metadata as a data frame
}
\examples{
read_metadata()
}
