% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs-spectrum.R
\name{read_pjnz_art_dec31}
\alias{read_pjnz_art_dec31}
\title{Read number on ART at Dec 31 from PJNZ}
\usage{
read_pjnz_art_dec31(pjnz)
}
\arguments{
\item{pjnz}{path to PJNZ file}
}
\description{
Reads the number on ART at December 31 from Spectrum PJNZ for children (0-14),
adult (15+) males, and adult (15+) females.
}
\examples{
pjnz <- system.file("extdata/demo_mwi2019.PJNZ", package = "naomi")
read_pjnz_art_dec31(pjnz)

}
