% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-data.R
\name{read_population}
\alias{read_population}
\alias{read_survey_indicators}
\alias{read_art_number}
\alias{read_anc_testing}
\alias{read_area_merged}
\title{Read Naomi structured input files}
\usage{
read_population(file)

read_survey_indicators(file)

read_art_number(file)

read_anc_testing(file)

read_area_merged(file)
}
\arguments{
\item{file}{A path to a file.}
}
\description{
Read Naomi structured input files
}
