% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-model.R
\name{read_rehydrate_zip}
\alias{read_rehydrate_zip}
\title{Read a rehydrate zip created by \code{save_rehydrate_zip}}
\usage{
read_rehydrate_zip(rehydrate_zip)
}
\arguments{
\item{rehydrate_zip}{The rehydrate zip to read}
}
\value{
The rehydrated zip as hintr_output object
}
\description{
This saves model output and plot data qs into a tempdir, and then
builds the hintr_output pointing to the tempdir
}
