% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-survey.R
\name{read_sf_zip_list}
\alias{read_sf_zip_list}
\title{Read Multiple Shape Files in ZIP Archive}
\usage{
read_sf_zip_list(zfile, pattern = "\\\\.shp$", read_fn = sf::read_sf)
}
\arguments{
\item{zfile}{path to a zip directory}

\item{pattern}{string pattern passed to \code{\link{list.files}}.}

\item{read_fn}{function used to read matched files.}
}
\description{
Reads all files in ZIP archive \code{zfile} matching \code{pattern} with
function \code{read_fn} and returns as a list.
}
