% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs-spectrum.R
\name{read_spectrum_region_code}
\alias{read_spectrum_region_code}
\title{Read Subnational Region Code from Spectrum PJNZ}
\usage{
read_spectrum_region_code(pjnz)
}
\arguments{
\item{pjnz}{file path to Spectrum PJNZ file.}
}
\value{
Spectrum subnational region code as an integer.
}
\description{
Read Subnational Region Code from Spectrum PJNZ
}
\details{
The region code is 0 if a national Spectrum file.
}
\examples{
pjnz <- system.file("extdata/demo_mwi2019.PJNZ", package = "naomi")
read_spectrum_region_code(pjnz)

}
