% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs-spectrum.R
\name{read_spectrum_region_name}
\alias{read_spectrum_region_name}
\title{Read Subnational Region Name from Spectrum PJNZ}
\usage{
read_spectrum_region_name(pjnz)
}
\arguments{
\item{pjnz}{file path to Spectrum PJNZ file.}
}
\value{
Spectrum subnational region name as a string. Returns NA
if no subnational region.
}
\description{
Read Subnational Region Name from Spectrum PJNZ
}
\details{
Value NA corresponds to region code 0 for a national Spectrum file.
}
\examples{
pjnz <- system.file("extdata/demo_mwi2019.PJNZ", package = "naomi")
read_spectrum_region_name(pjnz)

}
