% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{report_tmb}
\alias{report_tmb}
\title{Calculate Posterior Mean and Uncertainty Via TMB}
\usage{
report_tmb(naomi_fit)
}
\arguments{
\item{naomi_fit}{Fitted TMB model.}
}
\description{
Calculate Posterior Mean and Uncertainty Via TMB
}
