% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmb-model.R
\name{sample_tmb}
\alias{sample_tmb}
\title{Sample TMB fit}
\usage{
sample_tmb(fit, nsample = 1000, random_only = TRUE, verbose = TRUE)
}
\arguments{
\item{fit}{The TMB fit}

\item{nsample}{Number of samples}

\item{random_only}{Random only}

\item{verbose}{If TRUE prints additional information.}
}
\value{
Sampled fit.
}
\description{
Sample TMB fit
}
