% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputs.R
\name{save_output}
\alias{save_output}
\title{Save outputs to zip file}
\usage{
save_output(
  filename,
  dir,
  naomi_output,
  notes = NULL,
  vmmc_path = NULL,
  overwrite = FALSE,
  with_labels = FALSE,
  boundary_format = "geojson",
  single_csv = FALSE,
  export_datapack = !single_csv
)
}
\arguments{
\item{filename}{Name of file to create}

\item{dir}{Directory to create zip in}

\item{naomi_output}{Naomi output object}

\item{notes}{Notes to include in output zip}

\item{vmmc_path}{Path to VMMC excel workbook}

\item{overwrite}{If TRUE overwrite any existing file}

\item{with_labels}{If TRUE save indicator ids with labels}

\item{boundary_format}{Either geojson or shp for saving boundary as geojson
or shape format}

\item{single_csv}{If TRUE only output the csv of indicators, otherwise save
the metadata too}

\item{export_datapack}{If TRUE save CSV of PEPFAR datapack indicators.}

\item{options}{Naomi model options}
}
\value{
Path to created zip file
}
\description{
Save outputs to zip file
}
