% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputs.R
\name{save_output_package}
\alias{save_output_package}
\alias{read_output_package}
\title{Save outputs to zip file}
\usage{
save_output_package(
  naomi_output,
  filename,
  dir,
  overwrite = FALSE,
  with_labels = FALSE,
  boundary_format = "geojson",
  single_csv = FALSE,
  export_datapack = !single_csv
)

read_output_package(path)
}
\arguments{
\item{naomi_output}{Naomi output object}

\item{filename}{Name of file to create}

\item{dir}{Directory to create zip in}

\item{overwrite}{If TRUE overwrite any existing file}

\item{with_labels}{If TRUE save indicator ids with labels}

\item{boundary_format}{Either geojson or shp for saving boundary as geojson
or shape format}

\item{single_csv}{If TRUE only output the csv of indicators, otherwise save
the metadata too}

\item{export_datapack}{If TRUE save CSV of PEPFAR datapack indicators.}

\item{path}{Path to output zip file.}
}
\value{
Path to created zip file
}
\description{
Save outputs to zip file
}
