% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-model.R
\name{save_rehydrate_zip}
\alias{save_rehydrate_zip}
\title{Save out a zip file which can be loaded into project rehydrate}
\usage{
save_rehydrate_zip(hintr_output, path = tempfile(fileext = ".zip"))
}
\arguments{
\item{hintr_output}{The hintr_output object to save}

\item{path}{Path to save the file at}
}
\value{
Path to the newly saved file
}
\description{
Save out a zip file which can be loaded into project rehydrate
}
