% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/car.R
\name{scale_gmrf_precision}
\alias{scale_gmrf_precision}
\title{Scale of GMRF precision matrix}
\usage{
scale_gmrf_precision(
  Q,
  A = matrix(1, ncol = ncol(Q)),
  eps = sqrt(.Machine$double.eps)
)
}
\arguments{
\item{Q}{Precision matrix for a GMRF.}

\item{A}{Linear constraint for Q.}

\item{eps}{Value of the small constant added to the diagonal of Q for
invertibility.}
}
\description{
This function scales the precision matrix of a GMRF such that the geometric
mean of the marginal variance is one.
}
\details{
This implements the same thing as \code{\link[INLA:scale.model]{INLA::inla.scale.model}}. The marginal
variance of each connected component is one.
}
