% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report-plots.R
\name{scatter_plotly}
\alias{scatter_plotly}
\title{Plotly scatterplot data inputs and naomi outputs}
\usage{
scatter_plotly(
  df,
  ind,
  quarter,
  input_data,
  input_data_type,
  age_disag = "Y015_049",
  sex_disag = "both"
)
}
\arguments{
\item{df}{Inputs_outputs dataframe containing matched model estimates and data inputs.}

\item{ind}{Indicator filter.}

\item{quarter}{Calendar quarter filter.}

\item{input_data}{Input data name.}

\item{input_data_type}{Input data type.}

\item{age_disag}{Age group filter.}

\item{sex_disag}{Sex filter.}
}
\description{
Plotly scatterplot data inputs and naomi outputs
}
