% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmb-sdreport-jointprecision.R
\name{sdreport_joint_precision}
\alias{sdreport_joint_precision}
\title{Get Joint Precision of TMB Fixed and Random}
\usage{
sdreport_joint_precision(obj, par.fixed = NULL, hessian.fixed = NULL,
  bias.correct = FALSE, bias.correct.control = list(sd = FALSE, split =
  NULL, nsplit = NULL), ignore.parm.uncertainty = FALSE,
  skip.delta.method = FALSE)
}
\description{
This is a copy of \code{[TMB::sdreport]} that removes all computations of \code{ADREPORT()}ed variables to only return the joint precision.
}
