% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{select_naomi_data}
\alias{select_naomi_data}
\title{Select data for model fitting}
\usage{
select_naomi_data(naomi_mf, survey_hiv_indicators, anc_testing, art_number,
  prev_survey_ids, artcov_survey_ids, recent_survey_ids,
  vls_survey_ids = NULL, artnum_quarter_id_t1 = naomi_mf$quarter_id1,
  artnum_quarter_id_t2 = naomi_mf$quarter_id2,
  anc_prev_year_t1 = year_labels(naomi_mf$quarter_id1),
  anc_prev_year_t2 = year_labels(naomi_mf$quarter_id2),
  anc_artcov_year_t1 = anc_prev_year_t1,
  anc_artcov_year_t2 = anc_prev_year_t2)
}
\arguments{
\item{naomi_mf}{A Naomi model frame object.}

\item{survey_hiv_indicators}{Data frame of survey estimates, or NULL to exclude any survey data.}

\item{anc_testing}{Data frame of ANC routine testing outcomes, or NULL to exclude any ANC data.}

\item{art_number}{Data frame of number currently receiving ART, or NULL to exclude any ART data.}

\item{prev_survey_ids}{A character vector of \code{survey_id}s for prevalence data.}

\item{artcov_survey_ids}{A character vector of \code{survey_id}s for ART coverage data.}

\item{recent_survey_ids}{A character vector of \code{survey_id}s for recent HIV infection status.}

\item{vls_survey_ids}{A character vector of \code{survey_id}s for survey VLS among all HIV+ persons.}

\item{artnum_quarter_id_t1}{Quarter ID for first time point for number on ART.}

\item{artnum_quarter_id_t2}{Quarter ID for second time point for number on ART.}

\item{anc_quarter_id_t1}{Quarter IDs (possibly multiple) for first time point for ANC}

\item{anc_quarter_id_t2}{Quarter IDs (possibly multiple) for second time point for number on ART.}
}
\description{
Select data for model fitting
}
\details{
See example datasets for examples of required template for data sets. *\code{_survey_ids} must be reflected
in \code{survey_hiv_indicators}.

ART coverage and VLS survey data should not be included from the same survey. This is checked
by the function call and will throw an error.
}
\seealso{
\link{mwi_survey_hiv_indicators}, \link{mwi_anc_testing}, \link{mwi_art_number}, \link{convert_quarter_id}
}
