% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shipp.R
\name{shipp_adjust_sexbehav_fsw}
\alias{shipp_adjust_sexbehav_fsw}
\title{Adjust female sexual behavior risk groups by FSW proportions}
\usage{
shipp_adjust_sexbehav_fsw(outputs, options, fsw_est)
}
\arguments{
\item{outputs}{Naomi output.}

\item{options}{Naomi model options.}

\item{fsw_est}{5-year estimates of FSW PSEs generated from \code{shipp_disaggregate_fsw()}.

Estimates are generated for the following groups:
\itemize{
\item \code{nosex12m}:
\item \code{sexcohab}:
\item \code{sexnonregplus}:
\item \code{sexnonreg}:
\item \code{sexpaid12m}:
\item \code{nosex12m}:
}

Calculation steps:
\enumerate{
\item Align admin0/admin1 FSW proportions with SRB SAE estimates.
\item Subtract the proportion of FSW from total high risk female population.
}}
}
\value{
District level estimates of female sexual risk behaviour groups
}
\description{
Adjust female sexual behavior risk groups by FSW proportions
}
\keyword{internal}
