% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shipp.R
\name{shipp_adjust_sexbehav_msm_pwid}
\alias{shipp_adjust_sexbehav_msm_pwid}
\title{Adjust male sexual behavior risk groups by MSM + PWID proportions}
\usage{
shipp_adjust_sexbehav_msm_pwid(outputs, options, msm_est, pwid_est)
}
\arguments{
\item{outputs}{Naomi output.}

\item{options}{Naomi model options.}

\item{msm_est}{5-year estimates of MSM PSEs generated from \code{shipp__disaggregate_msm()}.}

\item{pwid_est}{5-year estimates of MSM PSEs generated from \code{shipp__disaggregate_pwid()}.}
}
\value{
District level estimates of male sexual risk behaviour groups
}
\description{
Adjust male sexual behavior risk groups by MSM + PWID proportions
}
\keyword{*}
\keyword{1.}
\keyword{2.}
\keyword{Align}
\keyword{Calculation}
\keyword{Estimates}
\keyword{MSM}
\keyword{PWID}
\keyword{SAE}
\keyword{SRB}
\keyword{Subtracting}
\keyword{`msm`:}
\keyword{`nosex12m`:}
\keyword{`pwid`:}
\keyword{`sexcohab`:}
\keyword{`sexonreg`:}
\keyword{`sexonregplus`:}
\keyword{admin0/admin1}
\keyword{all}
\keyword{and}
\keyword{are}
\keyword{estimates.}
\keyword{following}
\keyword{for}
\keyword{from}
\keyword{generated}
\keyword{groups.}
\keyword{groups:}
\keyword{internal}
\keyword{proportionally}
\keyword{proportions}
\keyword{steps:}
\keyword{the}
\keyword{with}
