% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shipp.R
\name{shipp_calculate_prevalence_female}
\alias{shipp_calculate_prevalence_female}
\title{Calculate prevalence for female SRB groups.}
\usage{
shipp_calculate_prevalence_female(
  naomi_output,
  options,
  fsw_est,
  female_srb,
  survey_year_sample = 2018
)
}
\arguments{
\item{naomi_output}{Naomi output.}

\item{options}{Naomi model options.}

\item{fsw_est}{5-year estimates of FSW PSEs generated from \code{shipp_disaggregate_fse()}.}

\item{female_srb}{FSW adjusted estimates of female SRB groups generated by \code{shipp_adjust_sexbehav_fsw()}}

\item{survey_year}{Year of survey to sample estimates.

To calculate district-age-sex-sexual behaviour-specific HIV prevalence, we maintain
HIV prevalence from Naomi for a district-age-sex, but disaggregate to different
risk behaviours using:
(1) HIV prevalence ratios from household surveys for those reporting no
sex vs one cohabiting vs non-regular sexual partner(s), and
(2) a linear regression through admin-1 level estimates of the ratio of KP
prevalence to gen-pop prevalence used to predict an age-district-specific
FSW to general population prevalence ratio.}
}
\value{
SRB PSEs with logit prevalence estimates.
}
\description{
Calculate prevalence for female SRB groups.
}
\keyword{internal}
