% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shipp.R
\name{shipp_calculate_prevalence_male}
\alias{shipp_calculate_prevalence_male}
\title{Calculate prevalence for male SRB groups.}
\usage{
shipp_calculate_prevalence_male(
  naomi_output,
  areas,
  options,
  msm_est,
  male_srb,
  survey_year_sample = 2018
)
}
\arguments{
\item{naomi_output}{Naomi output.}

\item{areas}{Naomi boundary file.}

\item{options}{Naomi model options.}

\item{msm_est}{5-year estimates of MSM PSEs generated from \code{shipp__disaggregate_msm()}.}

\item{male_srb}{MSM and PWID adjusted estimates of male SRB groups generated by \code{shipp_adjust_sexbehav_msm_pwid()}.}

\item{survey_year}{Year of survey to sample estimates.

To calculate district-age-sex-sexual behaviour-specific HIV prevalence, we maintain
HIV prevalence from Naomi for a district-age-sex, but disaggregate to different
risk behaviours using:
(1) HIV prevalence ratios from household surveys for those reporting no
sex vs one cohabiting vs non-regular sexual partner(s), and
(2) admin-1 level estimates of the ratio of KP prevalence to gen-pop prevalence
among 15-24 year olds for MSM (due to the young age distribution of MSM) or
among 15-49 year olds for PWID (due to the older age distribution of PWID)
applied to all age groups among MSM and PWID in districts by admin-1 unit.}
}
\value{
SRB PSEs with logit prevalence estimates.
}
\description{
Calculate prevalence for male SRB groups.
}
\keyword{internal}
