% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shipp.R
\name{shipp_disaggregate_fsw}
\alias{shipp_disaggregate_fsw}
\title{Dissagreggate admin1 FSW proportions from Oli's KP model to 5-age groups}
\usage{
shipp_disaggregate_fsw(outputs, options, naomi_pop, kp_consensus)
}
\arguments{
\item{outputs}{Naomi output.}

\item{options}{Naomi model options.}

\item{naomi_pop}{Naomi population estimates for T2.}

\item{kp_consensus}{Key pop consensus estimates.}
}
\value{
District level FSW estimates by 5-year age bands for ages 15-49.
}
\description{
Dissagreggate admin1 FSW proportions from Oli's KP model to 5-age groups
}
\keyword{internal}
