% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shipp.R
\name{shipp_format_naomi}
\alias{shipp_format_naomi}
\title{Format naomi outputs for PSE tool}
\usage{
shipp_format_naomi(outputs, options)
}
\arguments{
\item{outputs}{Naomi output}

\item{options}{Naomi model options.}
}
\value{
Naomi indicators formatted for the SHIPP workbook.
}
\description{
Format naomi outputs for PSE tool
}
\keyword{internal}
