% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-areas.R
\name{spread_areas}
\alias{spread_areas}
\title{Spread area hierarchy to wide format}
\usage{
spread_areas(
  areas,
  min_level = min(areas$area_level),
  max_level = max(areas$area_level)
)
}
\arguments{
\item{areas}{area hierarchy data.frame}

\item{min_level}{integer specifying the minimum level}

\item{max_level}{integer specifying the maximum level

TODO: Make this an example - where is areas.rds?
areas <- readRDS(system.file("extdata/areas/areas.rds", package = "naomi"))
areas_wide <- spread_areas(areas)}
}
\description{
Spread area hierarchy to wide format
}
