% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-areas.R
\name{st_read_zip}
\alias{st_read_zip}
\title{Read shape file from ZIP}
\usage{
st_read_zip(zfile, pattern = "shp$")
}
\arguments{
\item{zfile}{Path to zip file}

\item{pattern}{Pattern to read files for from zip, defaults to files ending
with 'shp'}
}
\description{
Read shape file from ZIP
}
