% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputs.R
\name{subset_naomi_output}
\alias{subset_naomi_output}
\title{Subset the results returned in Naomi output package}
\usage{
subset_naomi_output(
  naomi_output,
  area_id = NULL,
  area_level = NULL,
  sex = NULL,
  age_group = NULL,
  calendar_quarter = NULL,
  indicator = NULL,
  drop = FALSE,
  check_list = TRUE
)
}
\arguments{
\item{naomi_output}{Naomi output object.}

\item{area_id}{vector of area_ids to include/exclude.}

\item{area_level}{vector of area_levels to include/exclude.}

\item{sex}{vector of sexes to include/exclude.}

\item{age_group}{vector of age_groups to include/exclude.}

\item{calendar_quarter}{vector of calendar_quarters to include/exclude.}

\item{indicator}{vector of indicators to include/exclude.}

\item{drop}{logical whether to drop the supplied indices instead of keep
only the supplied indices (default).}

\item{check_list}{logical whether to check that supplied values are in the
output package to be subsetted.}
}
\value{
A naomi output package with a subset of results.
}
\description{
Subset the results returned in Naomi output package
}
\details{
If arguemnts are \code{NULL} (default), no subsetting is done on that dimension.

By default the argument \code{check_list = TRUE} means an error will be thrown
if any of the values in the vectors to subset are not found in the
\code{naomi_output} object supplied. This might be set to \code{FALSE} for some batch
processing applications, for example of the \code{naomi_output} could have already
been partially subsetted.
}
