% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputs.R
\name{subset_output_package}
\alias{subset_output_package}
\title{Resave a subsetted Naomi output package}
\usage{
subset_output_package(path, output_path, ...)
}
\arguments{
\item{path}{file path to naomi output package.}

\item{output_path}{path to resave subsetted output package.}

\item{...}{arguments to \code{\link[=subset_naomi_output]{subset_naomi_output()}}.}
}
\value{
path to saved output package.
}
\description{
This function reads an output package, subsets it using \code{\link[=subset_naomi_output]{subset_naomi_output()}}
and resaves the ouput package.
}
\details{
See \code{?subset_naomi_output()} for subsetting arguments and options.
}
\seealso{
\code{\link[=subset_naomi_output]{subset_naomi_output()}}
}
