% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{survey_prevalence_mf}
\alias{survey_prevalence_mf}
\title{Prepare model frames for survey datasets}
\usage{
survey_prevalence_mf(survey_ids, survey_hiv_indicators, naomi_mf)
}
\description{
Prepare model frames for survey datasets
}
