% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{tag_data_inputs}
\alias{tag_data_inputs}
\title{Tag data inputs}
\usage{
tag_data_inputs(full_data, filtered_data, cols = list())
}
\arguments{
\item{full_data}{Raw model input data}

\item{filtered_data}{Model input data filtered by model option specifications}
}
\value{
Tagged inputs
}
\description{
Tag data inputs
}
\keyword{internal}
