% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{update_mf_offsets}
\alias{update_mf_offsets}
\title{TODO: This should handle different age_max by sex...}
\usage{
update_mf_offsets(naomi_mf, prev_dat = NULL, artcov_dat = NULL,
  vls_dat = NULL)
}
\description{
TODO: This should handle different age_max by sex...
}
