% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmb-sdreport-jointprecision.R
\name{updateCholesky}
\alias{updateCholesky}
\alias{solveSubset}
\title{Non Exported Functions from TMB}
\usage{
updateCholesky(L, H, t = 0)

solveSubset(Q, L = Matrix::Cholesky(Q, super = TRUE, perm = TRUE),
  diag = FALSE)
}
\description{
Non Exported Functions from TMB
}
