% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-options.R
\name{validate_model_options}
\alias{validate_model_options}
\title{Validate a set of model options}
\usage{
validate_model_options(data, options)
}
\arguments{
\item{data}{The set of input data for the model run}

\item{options}{Key-value list of model options}
}
\value{
TRUE if valid otherwise throw an error
}
\description{
This validates that a set of model options can be used to run the model
}
