% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pepfar-datapack.R
\name{write_datapack_csv_from_pjnz}
\alias{write_datapack_csv_from_pjnz}
\title{Create National datapack CSV from PJNZ}
\usage{
write_datapack_csv_from_pjnz(
  pjnz,
  path = tempfile(fileext = ".csv"),
  psnu_level = 0
)
}
\arguments{
\item{pjnz}{file path to PJNZ file}
}
\value{
A data pack CSV file with outputs at national level
}
\description{
This constructs a dummy Naomi model fit at national level,
calibrates it to national Spectrum output, and writes
the datapack CSV.
}
\examples{
pjnz <- system.file("extdata/demo_mwi2024_v6.36.PJNZ", package = "naomi")
write_datapack_csv_from_pjnz(pjnz)

}
