% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unaids-navigator.R
\name{write_navigator_checklist}
\alias{write_navigator_checklist}
\title{Write UNAIDS Estimates Navigator checklist CSV}
\usage{
write_navigator_checklist(naomi_output, path)
}
\arguments{
\item{naomi_output}{a naomi_output object.}

\item{path}{path to save Navigator checklist CSV.}
}
\description{
Write UNAIDS Estimates Navigator checklist CSV
}
