% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{write_xlsx_sheets}
\alias{write_xlsx_sheets}
\title{Write list of data frames into an xlsx file}
\usage{
write_xlsx_sheets(template, sheets, path)
}
\arguments{
\item{template}{Path to xlsx file with sheets}

\item{sheets}{Named list of data frames to write into template. The names
must match the destination sheet in the xlsx}

\item{path}{Path to output the filled in xlsx}
}
\value{
Path to complete xlsx file
}
\description{
This doesn't write colmn headers into the workbook, it expects
that these already exist.
}
\keyword{internal}
