% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agyw-integration.R
\name{agyw_calculate_incidence_female}
\alias{agyw_calculate_incidence_female}
\title{Calculate incidence for female SRB groups.}
\usage{
agyw_calculate_incidence_female(
  naomi_output,
  options,
  female_srb,
  female_logit_prevalence,
  survey_year = 2018
)
}
\arguments{
\item{options}{Naomi options extracted from outputs}

\item{female_srb}{Estimates of female sexual risk groups generated by \code{agyw_adjust_sexbehav_fsw()}}

\item{female_logit_prevalence}{Risk adjusted estimates of female prevalence in sexual risk groups generated by \code{agyw_calculate_prevalence_female()}}

\item{survey_year}{Survey year to sample from the SAE model. Default is 2018.}

\item{outputs}{Naomi output.}
}
\value{
Wide format output required for the AGYW workbook.
}
\description{
While maintaining age/sex/district-specific HIV incidence from Naomi, distribute
HIV incidence between our 4 different behavioural groups utilizing IRRs from the
literature
}
\keyword{internal}
