% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agyw-integration.R
\name{agyw_calculate_incidence_male}
\alias{agyw_calculate_incidence_male}
\title{Calculate incidence in high risk male key populations}
\usage{
agyw_calculate_incidence_male(
  naomi_output,
  options,
  male_srb,
  male_logit_prevalence,
  survey_year = 2018
)
}
\arguments{
\item{options}{Naomi model options.}

\item{male_srb}{MSM and PWID adjusted estimated of male SRB groups generated by \code{agyw_adjust_sexbehav_msm_pwid()}.}

\item{male_logit_prevalence}{Risk adjusted estimates of male HIV prevalence in sexual risk groups generated by \code{agyw_calculate_prevalence_male()}.}

\item{survey_year}{Survey year to sample from the SAE model. Default is 2018. Survey year should be updated to most current household survey in the country - for countries without recent household surveys, leave at 2018 - the spatiotemporal
model of sexual behaviour fitted to all countries has the most data for in roughly 2018.}

\item{outputs}{Naomi output.}
}
\value{
Wide format output required for the AGYW workbook
}
\description{
Calculate incidence in high risk male key populations
}
