% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agyw-integration.R
\name{agyw_disaggregate_pwid}
\alias{agyw_disaggregate_pwid}
\title{Disaggregate admin1 PWID proportions from Oli's KP model to 5-age groups}
\usage{
agyw_disaggregate_pwid(outputs, options, naomi_pop, kp_consensus)
}
\arguments{
\item{outputs}{Naomi output.}

\item{options}{Naomi model options.}

\item{kp_consensus}{Key pop consensus estimates.}

\item{naomi_population}{Naomi population estimates for T2.}
}
\value{
District level PWID estimates by 5-year age bands for ages 15-49.
}
\description{
Disaggregate admin1 PWID proportions from Oli's KP model to 5-age groups
}
\keyword{internal}
