% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agyw-integration.R
\name{agyw_generate_risk_populations}
\alias{agyw_generate_risk_populations}
\title{Calculate incidence in high risk male key populations}
\usage{
agyw_generate_risk_populations(naomi_output, pjnz, survey_year = 2018)
}
\arguments{
\item{survey_year}{NOTES:: add in description + when this should be adjusted. Hardcoded to 2018.}

\item{outputs}{Naomi output.}

\item{options}{Naomi options extracted from outputs}

\item{male_srb}{Estimates of male sexual risk groups generated by \code{agyw_adjust_sexbehav_msm_pwid()}}

\item{male_logit_prevalence}{Risk adjusted estimates of male prevalence in sexual risk groups generated by \code{agyw_calculate_prevalence_male()}}
}
\value{
Wide format output required for the AGYW workbook
}
\description{
Survey year should be updated to most current household survey in the country -
for countries without recent household surveys, leave at 2018 - the spatiotemporal
model of sexual behaviour fitted to all countries has the most data for in roughly 2018
}
