% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{anc_testing_prev_mf}
\alias{anc_testing_prev_mf}
\alias{anc_testing_artcov_mf}
\alias{artnum_mf}
\title{Prepare Model Frames for Programme Datasets}
\usage{
anc_testing_prev_mf(year, anc_testing, naomi_mf)

anc_testing_artcov_mf(year, anc_testing, naomi_mf)

artnum_mf(quarter_id, art_number, naomi_mf)
}
\arguments{
\item{year}{Calendar year}

\item{anc_testing}{ART data frame}

\item{naomi_mf}{Naomi model frame}

\item{quarter_id}{quarter_id}

\item{art_number}{Number on ART}
}
\value{
Calculated prevalence
}
\description{
Prepare Model Frames for Programme Datasets
}
