% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{anc_testing_prev_mf}
\alias{anc_testing_prev_mf}
\alias{anc_testing_artcov_mf}
\alias{artnum_mf}
\title{Prepare Model Frames for Programme Datasets}
\usage{
anc_testing_prev_mf(quarter_ids, anc_testing, naomi_mf)

anc_testing_artcov_mf(quarter_ids, anc_testing, naomi_mf)

artnum_mf(quarter_id, art_number, naomi_mf)
}
\description{
Prepare Model Frames for Programme Datasets
}
