% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{anc_testing_prev_mf}
\alias{anc_testing_prev_mf}
\alias{anc_testing_artcov_mf}
\alias{anc_testing_clients_mf}
\title{Prepare Model Frames for Programme Datasets}
\usage{
anc_testing_prev_mf(year, anc_model_mf)

anc_testing_artcov_mf(year, anc_model_mf)

anc_testing_clients_mf(year, anc_model_mf)
}
\arguments{
\item{year}{Calendar year}

\item{anc_testing}{ART data frame}

\item{naomi_mf}{Naomi model frame}
}
\value{
Calculated prevalence
}
\description{
Prepare Model Frames for Programme Datasets
}
