% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agyw-integration.R
\name{assert_agyw_resource_hierarchy}
\alias{assert_agyw_resource_hierarchy}
\title{Throw warning when area hierarchy in external AGYW resources read in from
\code{naomi.resources} do not match Naomi outputs used to update AGYW estimates.}
\usage{
assert_agyw_resource_hierarchy(naomi_output, options)
}
\description{
@param naomi_output Naomi outputs.
@param options Naomi options.
}
