% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-survey.R
\name{calc_survey_hiv_indicators}
\alias{calc_survey_hiv_indicators}
\title{Calculate age/sex/area stratified survey estimates for biomarker outcomes}
\usage{
calc_survey_hiv_indicators(
  survey_meta,
  survey_regions,
  survey_clusters,
  survey_individuals,
  survey_biomarker,
  areas,
  sex = c("male", "female", "both"),
  age_group_id = NULL,
  area_top_level = min(areas$area_level),
  area_bottom_level = max(areas$area_level),
  artcov_definition = c("both", "arv", "artself"),
  by_restype = FALSE
)
}
\arguments{
\item{survey_meta}{Survey metadata.}

\item{survey_regions}{Survey regions.}

\item{survey_clusters}{Survey clusters.}

\item{survey_individuals}{Survey individuals.}

\item{survey_biomarker}{Survey biomarkers.}

\item{areas}{Areas.}

\item{sex}{Sex.}

\item{age_group_id}{Age group id.}

\item{area_top_level}{Area top level.}

\item{area_bottom_level}{Area bottom level.}

\item{artcov_definition}{Definition to use for calculate ART coverage.}

\item{by_restype}{Whether to stratify estimates by urban/rural restype; logical.}
}
\description{
Calculate age/sex/area stratified survey estimates for biomarker outcomes
}
\details{
All other data will be subsetted based on the \code{survey_id} values appearing in
survey_meta, so if only want to calculate for a subset of surveys it is
sufficient to pass subset for survey_meta and full data frames for the others.

Much of this function needs to be parsed out into more generic functions and
rewritten to be more efficient.
\itemize{
\item Age group would be more efficient if traversing a tree structure.
\item Need generic function to calculate
\item Flexibility about age/sex stratifications to calculate.
}

The argument \code{artcov_definition} controls whether to use both ARV biomarker and
self-report (\code{artcov_definition = "both"}; default), ARV biomarker only
(\code{artcov_definition = "arv"}), or self-report ART use only
(\code{artcov_definition = "artself"}).  If option is \code{"both"}, then all HIV positive
are used as the denomiator and no missing data on either indicator are
incorporated. If the option is \code{"arv"} or \code{"artself"} then missing values in those
variables, respectively, are treated as missing.
}
