% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shipp.R
\name{calculate_ywkp_pr_lor}
\alias{calculate_ywkp_pr_lor}
\title{Calculate YWKP prevalence ratio and log odds ratio}
\usage{
calculate_ywkp_pr_lor(prev, fit = ywkp_fit)
}
\arguments{
\item{prev}{(General population) prevalence}

\item{fit}{A model relating log-odds prevalence to YWKP log odds prevalence}
}
\description{
Calculate YWKP prevalence ratio and log odds ratio
}
\keyword{internal}
