% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-survey.R
\name{cmc_date}
\alias{cmc_date}
\title{Convert Date to Century Month Code (CMC)}
\usage{
cmc_date(date)
}
\arguments{
\item{date}{a Date vector}
}
\value{
an integer vector of CMC dates
}
\description{
Converts a date to DHS Century Month Code (CMC).
}
\details{
CMC date is defined as the number of months since 1900:
\deqn{cmc = (year - 1900) * 12 + momth}
}
\examples{
cmc_date(Sys.Date())
cmc_date(as.Date("1987-02-11", format = "\%Y-\%m-\%d"))

}
\references{
https://dhsprogram.com/Data/Guide-to-DHS-Statistics/Organization_of_DHS_Data.htm?rhtocid=_4_2_0#Structure_of_DHS_Databc-1
}
