% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/areas.R
\name{create_areas}
\alias{create_areas}
\title{Create an Areas Object}
\usage{
create_areas(levels, hierarchy, boundaries)
}
\arguments{
\item{levels}{Data frame of area level metadata.}

\item{hierarchy}{Data frame defining area hierarchy and area-level data.}

\item{boundaries}{an \code{sf} object with boundary geometry for each area_id}
}
\value{
An object of class \code{naomi_areas}
}
\description{
Constructs and validates an areas object as an S3 class.
}
\examples{
data(mwi_area_levels)
data(mwi_area_hierarchy)
data(mwi_area_boundaries)

areas <- create_areas(mwi_area_levels, mwi_area_hierarchy, mwi_area_boundaries)
areas

}
