% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/areas.R
\name{create_areas}
\alias{create_areas}
\title{Create an Areas Object}
\usage{
create_areas(
  levels = NULL,
  hierarchy = NULL,
  boundaries = NULL,
  area_merged = NULL
)
}
\arguments{
\item{levels}{Data frame of area level metadata.}

\item{hierarchy}{Data frame defining area hierarchy and area-level data.}

\item{boundaries}{an \code{sf} object with boundary geometry for each area_id}

\item{area_merged}{A merged version of \code{levels}, \code{hierarchy}, and \code{boundaries}.}
}
\value{
An object of class \code{naomi_areas}
}
\description{
Constructs and validates an areas object as an S3 class.
}
\examples{
data(demo_area_levels)
data(demo_area_hierarchy)
data(demo_area_boundaries)

areas <- create_areas(demo_area_levels, demo_area_hierarchy, demo_area_boundaries)
areas

}
