% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputs.R
\name{export_datapack}
\alias{export_datapack}
\title{Export naomi outputs to PEPFAR Data Pack format}
\usage{
export_datapack(
  naomi_output,
  path,
  psnu_level = max(naomi_output$meta_area$area_level),
  calendar_quarter = max(naomi_output$meta_period$calendar_quarter)
)
}
\arguments{
\item{naomi_output}{a naomi_output object.}

\item{path}{path to save Data Pack CSV.}

\item{psnu_level}{area_level for PEPFAR PSNU to export.}

\item{calendar_quarter}{calendar_quarter to export estimates.}
}
\description{
Export naomi outputs to PEPFAR Data Pack format
}
