% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs-spectrum.R
\name{extract_pjnz_naomi}
\alias{extract_pjnz_naomi}
\title{Extract Model Inputs from Spectrum PJNZ}
\usage{
extract_pjnz_naomi(pjnz_list)
}
\arguments{
\item{pjnz_list}{Vector of filepaths to Spectrum PJNZ file.}
}
\value{
A \code{data.frame} with Spectrum indicators.
}
\description{
Extract Model Inputs from Spectrum PJNZ
}
\examples{
pjnz <- system.file("extdata/demo_mwi2019.PJNZ", package = "naomi")
spec <- extract_pjnz_naomi(pjnz)

}
