% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmb-model.R
\name{fit_tmb}
\alias{fit_tmb}
\title{Fit TMB model}
\usage{
fit_tmb(
  tmb_input,
  outer_verbose = TRUE,
  inner_verbose = FALSE,
  max_iter = 250,
  progress = NULL
)
}
\arguments{
\item{tmb_input}{Model input data}

\item{outer_verbose}{If TRUE print function and parameters every iteration}

\item{inner_verbose}{If TRUE then disable tracing information from TMB}

\item{max_iter}{maximum number of iterations}

\item{progress}{Progress printer, if null no progress printed}
}
\value{
Fit model.
}
\description{
Fit TMB model
}
