% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{get_colour_scale}
\alias{get_colour_scale}
\title{Get colour scale for a particular indicator and country}
\usage{
get_colour_scale(indicator, country)
}
\arguments{
\item{indicator}{Indicator to get scale for}

\item{country}{Country to get scale for}
}
\value{
List of scale information including colour as a d3 scale chromatic
function name, whether to invert the scale and a min and max value for the
scale.
}
\description{
Get colour scale for a particular indicator and country
}
