% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{get_colour_scale}
\alias{get_colour_scale}
\title{Get colour scale and ranges for a particular country}
\usage{
get_colour_scale(iso3 = "default")
}
\arguments{
\item{iso3}{iso3 code of country to get metadata for or blank for default
configuration.}
}
\value{
List of scale information including colour as a d3 scale chromatic
function name, whether to invert the scale and a min and max value for the
scale.
}
\description{
Get colour scale and ranges for a particular country
}
\keyword{internal}
