% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-options.R
\name{get_model_options_template}
\alias{get_model_options_template}
\title{Get JSON template representing model run options}
\usage{
get_model_options_template(art, anc)
}
\arguments{
\item{art}{If FALSE then don't return template for ART control section.}

\item{anc}{If FALSE then don't return template for ANC control section.}
}
\value{
Model run options template.
}
\description{
This reads JSON file of model run options which describe how the Naomi
front end should display input options. Returns template for the API to
enrich with model options.
}
\examples{
get_model_options_template(TRUE, TRUE)
get_model_options_template(FALSE, FALSE)
}
