% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{get_plotting_metadata}
\alias{get_plotting_metadata}
\title{Get plotting metadata for a particular country}
\usage{
get_plotting_metadata(iso3)
}
\arguments{
\item{iso3}{iso3 code of country to get metadata for or blank for default
configuration.}
}
\value{
List of plotting metadata about how to locate data for a specific
indicator, data type and plot type. Also returns metadata about colour
scheme to use for that country and indicator.
}
\description{
Get plotting metadata for a particular country
}
\examples{
get_plotting_metadata("MWI")
}
