% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-model.R
\name{hintr_calibrate}
\alias{hintr_calibrate}
\title{Calibrate hintr_output}
\usage{
hintr_calibrate(
  output_path,
  calibration_options,
  plot_data_path = tempfile(fileext = ".rds"),
  calibrate_output_path = tempfile(fileext = ".rds")
)
}
\arguments{
\item{output_path}{A hintr_output object.}

\item{calibration_options}{A set of calibration options}

\item{plot_data_path}{Path to store calibrated output indicators as an RDS.}

\item{calibrate_output_path}{Path to store data required for re-calibrating model.}
}
\value{
Calibrated hintr_output object
}
\description{
Take a previously generated hintr_output object and calibrate. Format
response as another hintr_output object.
}
