% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloads.R
\name{hintr_prepare_datapack_download}
\alias{hintr_prepare_datapack_download}
\title{Prepare Datapack download}
\usage{
hintr_prepare_datapack_download(
  output,
  path = tempfile(fileext = ".xlsx"),
  vmmc_file = NULL,
  ids = NULL
)
}
\arguments{
\item{output}{hintr output object}

\item{path}{Path to save output file}

\item{vmmc_file}{Optional file object, with path, filename and hash for
VMMC input}

\item{ids}{List of naomi web app queue ids for putting into metadata}
}
\value{
Path to output file and metadata for file
}
\description{
Prepare Datapack download
}
