% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloads.R
\name{hintr_prepare_spectrum_download}
\alias{hintr_prepare_spectrum_download}
\title{Prepare spectrum download}
\usage{
hintr_prepare_spectrum_download(output, path = tempfile(fileext = ".zip"))
}
\arguments{
\item{path}{Path to save output file}

\item{hintr_output}{object}
}
\value{
Path to output file and metadata for file
}
\description{
Prepare spectrum download
}
