% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloads.R
\name{hintr_prepare_spectrum_download}
\alias{hintr_prepare_spectrum_download}
\title{Prepare spectrum download}
\usage{
hintr_prepare_spectrum_download(
  output,
  path = tempfile(fileext = ".zip"),
  notes = NULL
)
}
\arguments{
\item{output}{hintr output object}

\item{path}{Path to save output file}

\item{notes}{User added notes from front end of app as a string}
}
\value{
Path to output file and metadata for file
}
\description{
Prepare spectrum download
}
