% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-model.R
\name{hintr_run_model}
\alias{hintr_run_model}
\title{Run the model and save output to file}
\usage{
hintr_run_model(data, options, output_path = tempfile(),
  spectrum_path = tempfile(fileext = ".zip"),
  summary_path = tempfile(fileext = ".zip"))
}
\arguments{
\item{data}{List of paths to input data files. Must provide pjnz, shape,
population and survey data. Optionally include anc and art data too.}

\item{options}{List of model run options, must specify area_scope,
area_level, t1, t2, survey_prevalence, survey_art_coverage, survey_vls,
survey_recently_infected, survey_art_or_vls, art_t1, art_t2,
anc_prevalence_t1, anc_prevalence_t2, anc_art_coverage_t1,
anc_art_coverage_t2.}

\item{output_path}{Path to store output indicators as an RDS at.}

\item{spectrum_path}{Path to store spectrum digest file at.}

\item{summary_path}{Path to store summary download zip file at.}
}
\value{
Paths to 3 output files.
}
\description{
This prepares the model inputs from data and options and saves output as
an rds, the spectrum digst and indicators at specified paths.
}
