% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-model.R
\name{hintr_run_model}
\alias{hintr_run_model}
\title{Run the model and save output to file}
\usage{
hintr_run_model(
  data,
  options,
  output_path = tempfile(),
  spectrum_path = tempfile(fileext = ".zip"),
  coarse_output_path = tempfile(fileext = ".zip"),
  calibration_path = tempfile(fileext = ".rds")
)
}
\arguments{
\item{data}{List of paths to input data files.}

\item{options}{List of model run options (see details).}

\item{output_path}{Path to store output indicators as an RDS at.}

\item{spectrum_path}{Path to store spectrum digest file at.}

\item{coarse_output_path}{Path to store coarse age group output zip file at.}

\item{calibration_path}{Path to store data required for calibrating model.}
}
\value{
Paths to 4 output files.
}
\description{
This prepares the model inputs from data and options and saves output as
an rds, the spectrum digst and indicators at specified paths.
}
\details{
The \code{data} argument must be a list specifying paths to the following:
\itemize{
\item \code{pjnz}
\item \code{shape}
\item \code{population}
\item \verb{survey data}
\item \code{anc_testing} (optional)
\item \code{art_number} (optional)
}

Each item in list can either be a character containing the path to the file
or another list of the format:

list(
path = "path/to/file",
hash = "file_hash",
filename = "file"
)

The \code{options} argument must be a list specifying minimally:
\itemize{
\item area_scope
\item area_level
\item calendar_quarter_t1
\item calendar_quarter_t2
\item calendar_quarter_t3
\item survey_prevalence
\item survey_art_coverage
\item survey_recently_infected
\item include_art_t1
\item include_art_t2
\item anc_prevalence_year1
\item anc_prevalence_year2
\item anc_art_coverage_year1
\item anc_art_coverage_year2
}
}
