% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-population.R
\name{interpolate_population_agesex}
\alias{interpolate_population_agesex}
\title{Log-linear interpolation of age/sex stratified population}
\usage{
interpolate_population_agesex(population_agesex, quarter_ids)
}
\arguments{
\item{population_agesex}{a subset of the population_agesex.}

\item{quarter_id}{vector of quarter_ids to return interpolation.}
}
\value{
A data.frame with same columns as pop_agesex interpolated to \code{times}.
}
\description{
Log-linear interpolation of age/sex stratified population
}
\details{
\code{zoo::na.approx()} is used to interpolate log(population).
}
\examples{
## Interpolate Malawi population at level 2 (Zone) at two time points 
data(mwi_population_agesex)
quarter_ids <- convert_quarter_id(c(1, 3), c(2016, 2018))
pop_interp <- interpolate_population_agesex(mwi_population_agesex, quarter_ids)

}
\seealso{
\code{\link[=convert_quarter_id]{convert_quarter_id()}}
}
