% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-population.R
\name{interpolate_population_agesex}
\alias{interpolate_population_agesex}
\title{Log-linear interpolation of age/sex stratified population}
\usage{
interpolate_population_agesex(population_agesex, calendar_quarters)
}
\arguments{
\item{population_agesex}{a subset of the population_agesex.}

\item{calendar_quarters}{vector of calendar quarters to return interpolation.}
}
\value{
A data.frame with same columns as pop_agesex interpolated to \code{times}.
}
\description{
Log-linear interpolation of age/sex stratified population
}
\details{
\code{zoo::na.approx()} is used to interpolate log(population).
}
\examples{
## Interpolate Malawi population at level 2 (Zone) at two time points
data(demo_population_agesex)
calendar_quarters <- c("CY2016Q1", "CY2018Q3")
pop_interp <- interpolate_population_agesex(demo_population_agesex, calendar_quarters)

}
\seealso{
\code{\link[=convert_quarter_id]{convert_quarter_id()}}
}
