% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-population.R
\name{log_linear_interp}
\alias{log_linear_interp}
\title{Log-linear interpolation of NA values}
\usage{
log_linear_interp(y, x, rule = 2, replace_na = 0)
}
\arguments{
\item{y}{vector of output values, possibly with NAs}

\item{x}{vector of points to interpolate (no NAs)}

\item{rule}{rule for extrapolating outside range (see \code{\link[=approx]{approx()}})}

\item{replace_na}{value to replace if interpolation evaluates to NA}
}
\description{
Log-linear interpolation of NA values
}
\examples{
log_linear_interp(c(100, 105, NA, 110), 1:4)
log_linear_interp(c(NA, 105, NA, 110), 1:4)
log_linear_interp(c(NA, 105, NA, 110, NA), 1:5, rule = 1)
log_linear_interp(c(NA, 105, NA, 110, NA), 1:5, rule = 2)
log_linear_interp(c(NA, NA, 37), 1:3, rule = 2)

}
