% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report-plots.R
\name{map_outputs}
\alias{map_outputs}
\title{Create output map plot}
\usage{
map_outputs(
  naomi_geom,
  indicator,
  calendar_quarter,
  age,
  sex_disag,
  level = NULL,
  colour_palette = "red",
  fig_title = NULL,
  legend_title = NULL,
  legend_label = ggplot2::waiver(),
  breaks = 6
)
}
\arguments{
\item{naomi_geom}{Labelled naomi output object with geometry}

\item{indicator}{The indicator to generate plot for}

\item{calendar_quarter}{Naomi projection quarter to filter supplied data to.}

\item{age}{The age group to include in plots}

\item{sex_disag}{Sex diaggregation to use in plot,}

\item{level}{Integer, corresponding to desired area level to filter supplied data to. Default is lowest area level available in supplied data.}

\item{colour_palette}{Names or hexcode value for colour scale to be applied to to lowest and highest value in the supplied data. Can be specified as a preset colour palette using "red", "green" and "blue" or manually as a named list: cols = c( start_colour = "red", end_colour = "blue").}

\item{fig_title}{Title for the figure, blank by default}

\item{legend_title}{Title for the legend, blank by default}

\item{legend_label}{Legend label}

\item{breaks}{Number of break points to create in scale}
}
\value{
Mapped outputs
}
\description{
Create output map plot
}
