% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{naomi_model_frame}
\alias{naomi_model_frame}
\title{Construct Model Frames and Adjacency Structures}
\usage{
naomi_model_frame(area_merged, population_agesex, spec,
  scope = areas$tree$area_id,
  level = max(areas$tree$Get("area_level")), calendar_quarter1,
  calendar_quarter2, age_group_ids = 1:17, sexes = c("male", "female"),
  omega = 0.7, rita_param = list(OmegaT0 = 130/365, sigma_OmegaT =
  ((142 - 118)/365)/(2 * qnorm(0.975)), betaT0 = 0, sigma_betaT = 1e-05,
  ritaT = 1), sigma_u_sd = 1, artattend = FALSE,
  artattend_prior_sigma_scale = 3, logit_nu_mean = 2,
  logit_nu_sd = 0.3)
}
\arguments{
\item{area_merged}{Merged version of area hierarchy}

\item{population_agesex}{Population by age group and sex}

\item{spec}{Spec}

\item{scope}{The collection of area IDs to be modelled. Defaults to all area
ids.}

\item{level}{Admin level}

\item{calendar_quarter1}{Calendar quarter at time 1 ("CYyyyyQq")}

\item{calendar_quarter2}{Calendar quarter at time 2 ("CYyyyyQq")}

\item{age_group_ids}{Age group ids}

\item{sexes}{Sexes}

\item{omega}{Omega}

\item{rita_param}{rita_param}

\item{sigma_u_sd}{sigma_u_sd}

\item{artattend}{artattend}

\item{artattend_prior_sigma_scale}{artattend_propr_sigma_scale}

\item{logit_nu_mean}{mean of logit viral load suppression.}

\item{logit_nu_sd}{standard deviation of logit viral load suppression.}
}
\value{
Naomi model frame
}
\description{
Construct Model Frames and Adjacency Structures
}
