% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{naomi_model_frame}
\alias{naomi_model_frame}
\title{Construct Model Frames and Adjacency Structures}
\usage{
naomi_model_frame(
  area_merged,
  population_agesex,
  spec,
  scope = area_merged$area_id[area_merged$area_level == min(area_merged$area_level)],
  level = max(area_merged$area_level),
  calendar_quarter1,
  calendar_quarter2,
  calendar_quarter3,
  age_groups = get_five_year_age_groups(),
  sexes = c("male", "female"),
  omega = 0.7,
  rita_param = list(OmegaT0 = 130/365, sigma_OmegaT = ((142 - 118)/365)/(2 *
    qnorm(0.975)), betaT0 = 0, sigma_betaT = 1e-05, ritaT = 1),
  sigma_u_sd = 1,
  artattend = TRUE,
  artattend_t2 = FALSE,
  artattend_log_gamma_offset = -4,
  anchor_home_district = TRUE,
  rho_paed_15to49f_ratio = TRUE,
  rho_paed_x_term = FALSE,
  alpha_xst_term = FALSE,
  logit_nu_mean = 2,
  logit_nu_sd = 0.3,
  spectrum_population_calibration = "national",
  output_aware_plhiv = TRUE,
  adjust_area_growth = FALSE
)
}
\arguments{
\item{area_merged}{Merged version of area hierarchy}

\item{population_agesex}{Population by age group and sex}

\item{spec}{Spec}

\item{scope}{The collection of area IDs to be modelled. Defaults to all area
ids.}

\item{level}{Admin level}

\item{calendar_quarter1}{Calendar quarter at time 1 ("CYyyyyQq")}

\item{calendar_quarter2}{Calendar quarter at time 2 ("CYyyyyQq")}

\item{calendar_quarter3}{Calendar quarter at time 3 ("CYyyyyQq")}

\item{age_groups}{Age groups to include in model frame}

\item{sexes}{Sexes}

\item{omega}{Omega}

\item{rita_param}{rita_param}

\item{sigma_u_sd}{sigma_u_sd}

\item{artattend}{logical; whether to estimate neighboring district ART attendance}

\item{artattend_t2}{logical; whether to allow time-varying neighboring district ART attendance}

\item{artattend_log_gamma_offset}{logit offset for neighboring district ART attendance}

\item{anchor_home_district}{logical; whether to include home district random in ART attendance specification}

\item{rho_paed_15to49f_ratio}{logical; to model paediatric prevalence as ratio of 15-49 female prevalence}

\item{rho_paed_x_term}{logical; to include area interaction for paediatric prevalence}

\item{alpha_xst_term}{logical; to include district-sex-time interaction for ART coverage. Default \code{FALSE}.}

\item{logit_nu_mean}{mean of logit viral load suppression.}

\item{logit_nu_sd}{standard deviation of logit viral load suppression.}

\item{spectrum_population_calibration}{character string values "national", "subnational", "none"}

\item{adjust_area_growth}{TRUE/FALSE: adjust PLHIV population for net change in cohort size}
}
\value{
Naomi model frame
}
\description{
Construct Model Frames and Adjacency Structures
}
\details{
Argument \code{spectrum_population_calibration} determines whether to calibrate population
inputs to match Spectrum population by age and sex. If the Spectrum file is a single
national Spectrum file, then options "national" and "subnational" return the same results.
}
