% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{naomi_model_frame}
\alias{naomi_model_frame}
\title{Construct Model Frames and Adjacency Structures}
\usage{
naomi_model_frame(areas, population_agesex, spec, level, quarter_id1,
  quarter_id2, age_group_ids = 4:17, sexes = c("male", "female"))
}
\description{
Prune areas below model level
Model frame
Add population estimates
Add Spectrum inputs
Adjacency matrix
Scaled  precision matrix for 'BYM2' model.
Model output
}
