% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{naomi_output_frame}
\alias{naomi_output_frame}
\title{Model Frame and Linear Transform for Aggregated Model Outputs}
\usage{
naomi_output_frame(mf_model, areas, drop_partial_areas = TRUE)
}
\arguments{
\item{drop_partial_areas}{Drop areas from output if some children are
missing (default TRUE).}
}
\description{
Model Frame and Linear Transform for Aggregated Model Outputs
}
