% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputs.R
\name{output_package}
\alias{output_package}
\title{Build output package from fit}
\usage{
output_package(naomi_fit, naomi_mf)
}
\arguments{
\item{naomi_fit}{Fitted naomi model}

\item{naomi_mf}{Naomi model frame}
}
\value{
List containing output indicators and metadata.
}
\description{
Build output package from fit
}
