% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-time-series.R
\name{prepare_input_time_series_anc}
\alias{prepare_input_time_series_anc}
\title{Prepare data for ANC input time series plots}
\usage{
prepare_input_time_series_anc(anc, shape)
}
\arguments{
\item{anc}{Path to file containing ANC data or ANC data object}

\item{shape}{Path to file containing geojson areas data or shape sf object}
}
\value{
Data formatted for plotting input ANC time series containing columns
area_id, area_name, area_level, area_level_label, parent_area_id, area_sort_order,
time_period, year, quarter, calendar_quarter, area_hierarchy, plot, value and
missing_ids
}
\description{
Take uploaded ANC and shape file paths or objects and format as data which
can be used to draw input time series graphs.
}
