% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmb-model.R
\name{prepare_tmb_inputs}
\alias{prepare_tmb_inputs}
\title{Prepare inputs for TMB model.}
\usage{
prepare_tmb_inputs(naomi_data, report_likelihood = 1L)
}
\arguments{
\item{naomi_data}{Naomi data object}

\item{report_likelihood}{Option to report likelihood in fit object (default true).}

\item{anchor_home_district}{Option to include random effect home district attractiveness to retain residents on ART within home districts (default true).}
}
\value{
Inputs ready for TMB model
}
\description{
Prepare inputs for TMB model.
}
\seealso{
\link{select_naomi_data}
}
