% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmb-model.R
\name{prepare_tmb_inputs}
\alias{prepare_tmb_inputs}
\title{ANC prevalence model matrices
ART attendance aggregation}
\usage{
prepare_tmb_inputs(naomi_mf, prev_dat, artcov_dat, recent_dat,
  anc_prev_t1_dat, anc_prev_t2_dat, anc_artcov_t1_dat, anc_artcov_t2_dat,
  artnum_t1_dat, artnum_t2_dat)
}
\description{
ANC prevalence model matrices
ART attendance aggregation
}
