% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmb-model.R
\name{prepare_tmb_inputs}
\alias{prepare_tmb_inputs}
\title{Prepare inputs for TMB model.}
\usage{
prepare_tmb_inputs(naomi_data)
}
\arguments{
\item{naomi_data}{Naomi data object}
}
\value{
Inputs ready for TMB model
}
\description{
Prepare inputs for TMB model.
}
\seealso{
\link{select_naomi_data}
}
