% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-population.R
\name{quarter_year_labels}
\alias{quarter_year_labels}
\alias{quarter_number}
\alias{quarter_labels}
\alias{calendar_quarter_labels}
\alias{year_labels}
\alias{convert_quarter_id}
\alias{convert_calendar_quarter}
\alias{calendar_quarter_to_quarter_id}
\alias{quarter_id_to_calendar_quarter}
\alias{calendar_quarter_to_year}
\title{Time period indexing}
\usage{
quarter_year_labels(quarter_id)

quarter_number(quarter_id)

quarter_labels(quarter_id)

calendar_quarter_labels(calendar_quarter)

year_labels(quarter_id)

convert_quarter_id(year, quarter)

convert_calendar_quarter(year, quarter)

calendar_quarter_to_quarter_id(calendar_quarter)

quarter_id_to_calendar_quarter(quarter_id)

calendar_quarter_to_year(calendar_quarter)
}
\arguments{
\item{quarter_id}{vector of integer quarter IDs.}

\item{calendar_quarter}{Vector of calendar quarters to convert.}

\item{year}{vector of integer years.}

\item{quarter}{vector of integer quarters (1,2,3,4).}
}
\description{
Time periods are indexed by integers for efficiency and precision.
Quarters are indexed as the number of quarters since the beginning of
1900: $quarter_id = (year - 1900) * 4 + quarter$.
}
\details{
Quarters are labelled as "Jan-Mar", "Apr-Jun", "Jul-Sep", "Oct-Dec" instead of
"Q1", "Q2", "Q3", "Q4" to avoid confusion between calendar quarters and offset
fiscal year quarters.
}
\examples{
quarter_ids <- convert_quarter_id(c(2009, 2017), c(3, 1))
quarter_ids
calender_quarters <- convert_calendar_quarter(c(2009, 2017), c(3, 1))
quarter_number(quarter_ids)
quarter_labels(quarter_ids)
year_labels(quarter_ids)
quarter_year_labels(quarter_ids)
calendar_quarter_labels("CY2015Q2")
calendar_quarter_to_year("CY2015Q2")

}
