% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmb-model.R
\name{sample_tmb}
\alias{sample_tmb}
\title{Sample from Joint Posterior Distribution}
\usage{
sample_tmb(fit, nsample = 1000, random_only = TRUE, verbose = TRUE)
}
\description{
Sample from Joint Posterior Distribution
}
