% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmb-sdreport-jointprecision.R
\name{sdreport_joint_precision}
\alias{sdreport_joint_precision}
\title{Get Joint Precision of TMB Fixed and Random}
\usage{
sdreport_joint_precision(
  obj,
  par.fixed = NULL,
  hessian.fixed = NULL,
  bias.correct = FALSE,
  bias.correct.control = list(sd = FALSE, split = NULL, nsplit = NULL),
  ignore.parm.uncertainty = FALSE,
  skip.delta.method = FALSE
)
}
\arguments{
\item{obj}{Object returned by \code{MakeADFun}}

\item{par.fixed}{Optional. Parameter estimate (will be known to \code{obj} when an optimization has been carried out).}

\item{hessian.fixed}{Optional. Hessian wrt. parameters (will be calculated from \code{obj} if missing).}

\item{bias.correct}{logical indicating if bias correction should be applied}

\item{bias.correct.control}{a \code{list} of bias correction options; currently \code{sd}, \code{split} and \code{nsplit} are used - see details.}

\item{ignore.parm.uncertainty}{Optional. Ignore estimation variance of parameters?}

\item{skip.delta.method}{Skip the delta method? (\code{FALSE} by default)}
}
\description{
This is a copy of \verb{[TMB::sdreport]} that removes all computations of \code{ADREPORT()}ed variables to only return the joint precision.
}
\keyword{internal}
