% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{survey_mf}
\alias{survey_mf}
\title{Prepare model frames for survey datasets}
\usage{
survey_mf(
  survey_ids,
  indicator,
  survey_hiv_indicators,
  naomi_mf,
  use_kish = TRUE,
  deff = 1,
  min_age = 0,
  max_age = 80,
  use_aggregate = FALSE
)
}
\arguments{
\item{survey_ids}{Survey IDs}

\item{indicator}{Indicator to filter, character string}

\item{survey_hiv_indicators}{Survey HIV indicators}

\item{naomi_mf}{Naomi model frame}

\item{use_kish}{Logical whether to use Kish effective sample size}

\item{deff}{Assumed design effect for scaling effective sample size}

\item{min_age}{Min age for calculating recent infection}

\item{max_age}{Max age for calculating recent infection}

\item{use_aggregate}{Logical; use aggregate survey data as provided instead of
subsetting fine area/sex/age group stratification.}
}
\description{
Prepare model frames for survey datasets
}
